Data.CBytes

Documentation

data CBytes

A efficient wrapper for null-terminated string which can be automatically freed by ghc garbage collector.

The main design target of this type is to ease the bridging of C FFI APIs, since most of the unix APIs use null-terminated string. On windows you're encouraged to use a compatibility layer like 'WideCharToMultiByte/MultiByteToWideChar' and keep the same interface.

We neither guarantee to store length info, nor support O(1) slice for CBytes: This will defeat the purpose of null-terminated string which is to save memory, and strlen runs very fast.

It can be used with OverloadedString, literal encoding is UTF-8 with some modifications: NUL char is encoded to 'C0 80', and '\xD800' ~ '\xDFFF' is encoded as a three bytes normal utf-8 codepoint. This is also how ghc compile string literal into binaries, thus we can use rewrite-rules to construct CBytes value in O(1) without wasting runtime heap.

Note most of the unix API is not unicode awared though, you may find a scandir call return a filename which is not proper encoded in any unicode encoding at all. But still, UTF-8 is recommanded to be used everywhere, so we use that assumption in various places, such as displaying CBytes and converting literals.

Constructors

CBytesOnHeap !(MutablePrimArray RealWorld Word8)
On heap pinned MutablePrimArray

CBytesLiteral !CString
String literals with static address

packCBytes :: String -> CBytes

Pack a String into null-terminated CByte.

data SP2

Constructors

SP2 !Int !(MutablePrimArray RealWorld Word8)

unpackCBytes :: CBytes -> String

fromCStringMaybe :: HasCallStack => CString -> IO (Maybe CBytes)

Copy a CString type into a CBytes, return Nothing if the pointer is NULL.

After copying you're free to free the CString 's memory.

fromCString :: HasCallStack => CString -> IO CBytes

Same with fromCString, but throw InvalidArgument when meet a null pointer.

withCBytes :: CBytes -> (CString -> IO a) -> IO a

Pass CBytes to foreign function as a char*.

c_strcmp :: CString -> CString -> IO CInt

c_strlen :: CString -> IO CSize

c_strcpy :: MutableByteArray# RealWorld -> CString -> IO ()