Data.Primitive.PrimArray

Description

This module provide primitive arrays tagged with element type, all offset, length .. ,etc. is based on elemment not byte, and operations are NOT bound checked.

You can define new primitive array by defining Prim instances, but since GHC heap array are word aligned, It's not recommand to use PrimArray with a type which alignment is not submultiple of word size. Otherwise you have to use newAlignedPinnedPrimArray to make a new array for such a type.

Here is an example of a RGB pixel Prim instance.

  data Pixel = Pixel Word8 Word8 Word8 -- you may want to unpack these
  instance Prim Pixel where
    sizeOf
    alignment
    indexByteArray i ba)
                                (indexByteArray (i 1#))
                                (indexByteArray (i 2#))
    ...

Now you can use 'PrimArray Pixel' with either this module or "Data.Array/Data.Vector".

If you're trying to do FFI with primitive array, please use functions from Foreign.PrimArray instead of using withPrimArrayContent / withMutablePrimArrayContents directly, Read notes in Foreign.PrimArray to understand why these functions are dangerous.

Types

newtype PrimArray a

Primitive array tagged with element type a.

Constructors

PrimArray ByteArray

newtype MutablePrimArray s a

Mutable primitive array tagged with element type a.

Constructors

MutablePrimArray (MutableByteArray s)

Allocation

newPrimArray :: forall m a. (PrimMonad m, Prim a) => Int -> m (MutablePrimArray (PrimState m) a)

Create a new mutable primitive array of the specified size.

newPinnedPrimArray :: forall m a. (PrimMonad m, Prim a) => Int -> m (MutablePrimArray (PrimState m) a)

Create a pinned byte array of the specified size and respect the primitive type's alignment. The garbage collector is guaranteed not to move it.

newAlignedPinnedPrimArray :: forall m a. (PrimMonad m, Prim a) => Int -> m (MutablePrimArray (PrimState m) a)

Create a pinned primitive array of the specified size and respect given primitive type's alignment. The garbage collector is guaranteed not to move it.

Element access

readPrimArray :: forall m a. (PrimMonad m, Prim a) => MutablePrimArray (PrimState m) a -> Int -> m a

Read a primitive value from the primitive array. The offset is given in elements of type a.

writePrimArray :: forall m a. (PrimMonad m, Prim a) => MutablePrimArray (PrimState m) a -> Int -> a -> m ()

Write a primitive value to the primitive array. The offset is given in elements of type a.

indexPrimArray :: Prim a => PrimArray a -> Int -> a

Read a primitive value from the primitive array. The offset is given in elements of type a.

Freezing and thawing

unsafeFreezePrimArray :: PrimMonad m => MutablePrimArray (PrimState m) a -> m (PrimArray a)

Convert a mutable primitive array to an immutable one without copying. The array should not be modified after the conversion.

unsafeThawPrimArray :: PrimMonad m => PrimArray a -> m (MutablePrimArray (PrimState m) a)

Convert an immutable primitive array to a mutable one without copying. The original array should not be used after the conversion.

Block operations

copyPrimArray

Arguments

:: (PrimMonad m, Prim a)

=> MutablePrimArray (PrimState m) a
destination array

-> Int
offset into destination array

-> PrimArray a
source array

-> Int
offset into source array

-> Int
number of prims to copy

-> m ()

Copy a slice of an immutable primitive array to a mutable primitive array. The offset and length are given in elements of type a.

copyMutablePrimArray

Arguments

:: (PrimMonad m, Prim a)

=> MutablePrimArray (PrimState m) a
destination array

-> Int
offset into destination array

-> MutablePrimArray (PrimState m) a
source array

-> Int
offset into source array

-> Int
number of prims to copy

-> m ()

Copy a slice of a mutable primitive array into another array. The two slices may not overlap. The offset and length are given in elements of type a.

movePrimArray

Arguments

:: (PrimMonad m, Prim a)

=> MutablePrimArray (PrimState m) a
destination array

-> Int
offset into destination array

-> MutablePrimArray (PrimState m) a
source array

-> Int
offset into source array

-> Int
number of prims to copy

-> m ()

Copy a slice of a mutable primitive array into another, potentially overlapping array. The offset and length are given in elements of type a.

setPrimArray

Arguments

:: (PrimMonad m, Prim a)

=> MutablePrimArray (PrimState m) a
array to fill

-> Int
offset into array

-> Int
number of values to fill

-> a
value to fill with

-> m ()

Fill a slice of a mutable primitive array with a value. The offset and length are given in elements of type a.

resizeMutablePrimArray :: forall m a. (PrimMonad m, Prim a) => MutablePrimArray (PrimState m) a -> Int -> m (MutablePrimArray (PrimState m) a)

Resize a primitive array using resizeMutableByteArray#.

To avoid undefined behaviour, the original MutablePrimArray shall not be accessed anymore.

shrinkMutablePrimArray :: forall m a. (Prim a, PrimMonad m) => MutablePrimArray (PrimState m) a -> Int -> m ()

Shrink a primitive array using shrinkMutableByteArray#.

The new size argument must be less than or equal to the current size, but it's not checked.

copyPrimArrayToPtr

Arguments

:: (PrimMonad m, Prim a)

=> Ptr a
destination pointer

-> PrimArray a
source array

-> Int
offset into source array

-> Int
number of prims to copy

-> m ()

Copy a slice of an immutable primitive array to an address. The offset and length are given in elements of type a.

copyMutablePrimArrayToPtr

Arguments

:: (PrimMonad m, Prim a)

=> Ptr a
destination pointer

-> MutablePrimArray (PrimState m) a
source array

-> Int
offset into source array

-> Int
number of prims to copy

-> m ()

Copy a slice of an mutable primitive array to an address. The offset and length are given in elements of type a.

copyMutablePrimArrayFromPtr

Arguments

:: (PrimMonad m, Prim a)

=> MutablePrimArray (PrimState m) a
destination array

-> Int
offset into destination array

-> Ptr a
source pointer

-> Int
number of prims to copy

-> m ()

Copy a slice of an mutable primitive array from an address. The offset and length are given in elements of type a.

Information

sizeofPrimArray :: forall a. Prim a => PrimArray a -> Int

Size of the primitive array.

sizeofMutablePrimArray :: forall m a. (PrimMonad m, Prim a) => MutablePrimArray (PrimState m) a -> m Int

Size of the mutable primitive array.

sameMutablePrimArray :: MutablePrimArray s a -> MutablePrimArray s a -> Bool

Check if the two arrays refer to the same memory block.

primArrayContents :: PrimArray a -> Ptr a

Yield a pointer to the array's data. This operation is only safe on pinned primitive arrays allocated by newPinnedPrimArray or newAlignedPinnedPrimArray, and you have to make sure the PrimArray can outlive the Ptr.

mutablePrimArrayContents :: MutablePrimArray s a -> Ptr a

Yield a pointer to the array's data.

This operation is only safe on pinned primitive arrays allocated by newPinnedPrimArray or newAlignedPinnedPrimArray. and you have to make sure the PrimArray can outlive the Ptr.

withPrimArrayContents :: PrimMonad m => PrimArray a -> (Ptr a -> m b) -> m b

Yield a pointer to the array's data and do computation with it.

This operation is only safe on pinned primitive arrays allocated by newPinnedPrimArray or newAlignedPinnedPrimArray.

withMutablePrimArrayContents :: PrimMonad m => MutablePrimArray (PrimState m) a -> (Ptr a -> m b) -> m b

Yield a pointer to the array's data and do computation with it.

This operation is only safe on pinned primitive arrays allocated by newPinnedPrimArray or newAlignedPinnedPrimArray.

isPrimArrayPinned :: PrimArray a -> Bool

Check if a primitive array is pinned.

isMutablePrimArrayPinned :: MutablePrimArray s a -> Bool

Check if a mutable primitive array is pinned.

samePrimArray :: PrimArray a -> PrimArray a -> Bool

Check if the two immutable arrays refer to the same memory block.

Prefetch

Check http://hackage.haskell.org/package/ghc-prim-0.3.1.0/docs/GHC-Prim.html#Prefetch for more information.

prefetchPrimArray0 :: PrimMonad m => PrimArray a -> Int -> m ()

prefetchPrimArray1 :: PrimMonad m => PrimArray a -> Int -> m ()

prefetchPrimArray2 :: PrimMonad m => PrimArray a -> Int -> m ()

prefetchPrimArray3 :: PrimMonad m => PrimArray a -> Int -> m ()

prefetchMutablePrimArray0 :: PrimMonad m => MutablePrimArray (PrimState m) a -> Int -> m ()

prefetchMutablePrimArray1 :: PrimMonad m => MutablePrimArray (PrimState m) a -> Int -> m ()

prefetchMutablePrimArray2 :: PrimMonad m => MutablePrimArray (PrimState m) a -> Int -> m ()

prefetchMutablePrimArray3 :: PrimMonad m => MutablePrimArray (PrimState m) a -> Int -> m ()