Data.Text

Documentation

newtype Text

Text represented as UTF-8 encoded Bytes

Constructors

Text

Fields

toUTF8Bytes :: Bytes

compareText :: Text -> Text -> Ordering

data DecodeResult

Constructors

Success !Text

PartialBytes !Text !Bytes

InvalidBytes !Bytes

validateUTF8 :: Bytes -> DecodeResult

validateUTF8_ :: Bytes -> (Text, Bytes)

repairUTF8 :: Bytes -> (Text, Bytes)

https://stackoverflow.com/questions/2547262/why-is-python-decode-replacing-more-than-the-invalid-bytes-from-an-encoded-strin

pack :: String -> Text

packN :: Int -> String -> Text

O(n) Convert a list into a text with an approximate size(in bytes, not codepoints).

If the encoded bytes length is larger than the size given, we simply double the buffer size and continue building.

This function is a good consumer in the sense of build/foldr fusion.

data SP2 s

Constructors

SP2 !Int !(MutablePrimArray s Word8)

unpack :: Text -> String

unpackFB :: Text -> (Char -> a -> a) -> a -> a

unpackR :: Text -> String

unpackRFB :: Text -> (Char -> a -> a) -> a -> a

singleton :: Char -> Text

empty :: Text

Basic interface

cons :: Char -> Text -> Text

snoc :: Text -> Char -> Text

append :: Text -> Text -> Text

uncons :: Text -> Maybe (Char, Text)

unsnoc :: Text -> Maybe (Text, Char)

head :: Text -> Char

tail :: Text -> Text

last :: Text -> Char

init :: Text -> Text

null :: Text -> Bool

length :: Text -> Int

Transformations

map :: (Char -> Char) -> Text -> Text

O(n) map f t is the Text obtained by applying f to each element of t. Performs replacement on invalid scalar values.

intercalate :: Text -> [Text] -> Text

O(n) The intercalate function takes a Text and a list of Texts and concatenates the list after interspersing the first argument between each element of the list.

concat :: [Text] -> Text

intersperse :: Char -> Text -> Text

O(n) The intersperse function takes a character and places it between the characters of a Text. Performs replacement on invalid scalar values.

reverse :: Text -> Text

O(n) Reverse the characters of a string.

(/../) :: Text -> (Int, Int) -> Text

errorEmptyText :: String -> a