Data.Vector

Description

This module provide fast boxed and unboxed vector with unified interface. Conceptually a vector is simply a slice of an array, for example this is the definition of boxed vector:

    data Vector a = Vector
         !(SmallArray a) -- payload
         !Int         -- offset
         !Int         -- length

The Vec class unified different type of vectors, and this module provide operation over Vec instances. The API is a combination of bytestring, text and vector. If you find missing functions, please report!

Vec typeclass

class Arr (MArray v) (IArray v) a => Vec v a where

Typeclass for box and unboxed vectors, which are created by slicing arrays.

Minimal complete definition

toArr, fromArr

Associated Types

type MArray v :: * -> * -> *

Vector's mutable array type

type IArray v :: * -> *

Vector's immutable array type

Methods

toArr :: v a -> (IArray v a, Int, Int)

Get underline array and slice range(offset and length).

fromArr :: IArray v a -> Int -> Int -> v a

Create a vector by slicing an array(with offset and length).

Boxed and unboxed vector type

data Vector a

Boxed vector

Constructors

Vector !(SmallArray a) !Int !Int

pattern VecPat :: forall (v :: * -> *) a. Vec v a => IArray v a -> Int -> Int -> v a

A pattern synonyms for matching the underline array, offset and length.

data PrimVector a

Primitive vector

Constructors

PrimVector !(PrimArray a) !Int !Int

vW16 :: QuasiQuoter

Word8 vector

type Bytes = PrimVector Word8

Bytes is just primitive word8 vectors.

pattern BytesPat :: forall (v :: * -> *) a. ((~#) (* -> *) (* -> *) (IArray v) PrimArray, Vec v a) => ByteArray -> Int -> Int -> v a

vASCII :: QuasiQuoter

w2c :: Word8 -> Char

Conversion between Word8 and Char. Should compile to a no-op.

c2w :: Char -> Word8

Unsafe conversion between Char and Word8. This is a no-op and silently truncates to 8 bits Chars > '\255'. It is provided as convenience for PrimVector construction.

Basic creating

create :: Vec v a => Int -> (forall s. MArray v s a -> ST s ()) -> v a

Create a vector.

creating :: Vec v a => Int -> (forall s. MArray v s a -> ST s b) -> (b, v a)

Create a vector, return both the vector and the monadic result during creating.

createN :: Vec v a => Int -> (forall s. MArray v s a -> ST s Int) -> v a

Create a vector up to a specific length.

If the initialization function return a length larger than initial size, an error will be raised.

empty :: Vec v a => v a

O(1). The empty vector.

singleton :: Vec v a => a -> v a

O(1). Single element vector.

copy :: Vec v a => v a -> v a

O(n). Copy a vector from slice.

Conversion between list

pack :: Vec v a => [a] -> v a

O(n) Convert a list into a vector

Alias for packN defaultInitSize.

packN :: forall v a. Vec v a => Int -> [a] -> v a

O(n) Convert a list into a vector with an approximate size.

If the list's length is large than the size given, we simply double the buffer size and continue building.

This function is a good consumer in the sense of build/foldr fusion.

packR :: Vec v a => [a] -> v a

O(n)

Alias for packRN defaultInitSize.

packRN :: forall v a. Vec v a => Int -> [a] -> v a

O(n) packN in reverse order.

This function is a good consumer in the sense of build/foldr fusion.

unpack :: Vec v a => v a -> [a]

O(n) Convert vector to a list.

Unpacking is done lazily. i.e. we will retain reference to the array until all element are consumed.

This function is a good producer in the sense of build/foldr fusion.

unpackR :: Vec v a => v a -> [a]

O(n) Convert vector to a list in reverse order.

This function is a good producer in the sense of build/foldr fusion.

Basic interface

append :: Vec v a => v a -> v a -> v a

O(m+n)

There's no need to guard empty vector because we guard them for you, so appending empty vectors are no-ops.

null :: Vec v a => v a -> Bool

O(1) Test whether a vector is empty.

length :: Vec v a => v a -> Int

O(1) The length of a vector.

cons :: Vec v a => a -> v a -> v a

O(n) cons is analogous to (:) for lists, but of different complexity, as it requires making a copy.

snoc :: Vec v a => v a -> a -> v a

O(n) Append a byte to the end of a vector

uncons :: Vec v a => v a -> Maybe (a, v a)

O(1) Extract the head and tail of a PrimVector, returning Nothing if it is empty.

unsnoc :: Vec v a => v a -> Maybe (v a, a)

O(1) Extract the init and last of a PrimVector, returning Nothing if it is empty.

head :: Vec v a => v a -> a

O(1) Extract the first element of a PrimVector, which must be non-empty. An exception will be thrown in the case of an empty PrimVector.

tail :: Vec v a => v a -> v a

O(1) Extract the elements after the head of a PrimVector, which must be non-empty. An exception will be thrown in the case of an empty PrimVector.

last :: Vec v a => v a -> a

O(1) Extract the first element of a PrimVector, which must be non-empty. An exception will be thrown in the case of an empty PrimVector.

init :: Vec v a => v a -> v a

O(1) Extract the elements after the head of a PrimVector, which must be non-empty. An exception will be thrown in the case of an empty PrimVector.

Transforming vector

map :: forall u v a b. (Vec u a, Vec v b) => (a -> b) -> u a -> v b

reverse :: forall v a. Vec v a => v a -> v a

O(n) reverse xs efficiently returns the elements of xs in reverse order.

intersperse :: forall v a. Vec v a => a -> v a -> v a

O(n) The intersperse function takes an element and a vector and `intersperses' that element between the elements of the vector. It is analogous to the intersperse function on Lists.

intercalate :: Vec v a => v a -> [v a] -> v a

O(n) The intercalate function takes a vector and a list of vectors and concatenates the list after interspersing the first argument between each element of the list.

Note: intercalate will force the entire vector list.

intercalateElem :: Vec v a => a -> [v a] -> v a

O(n) An efficient way to join vector with an element.

transpose :: Vec v a => [v a] -> [v a]

The transpose function transposes the rows and columns of its vector argument.

Reducing vector (folds)

foldl' :: Vec v a => (b -> a -> b) -> b -> v a -> b

foldl1' :: forall v a. Vec v a => (a -> a -> a) -> v a -> a

foldr' :: Vec v a => (a -> b -> b) -> b -> v a -> b

foldr1' :: forall v a. Vec v a => (a -> a -> a) -> v a -> a

Special folds

concat :: forall v a. Vec v a => [v a] -> v a

O(n) Concatenate a list of vector.

Note: concat have to force the entire list to filter out empty vector and calculate the length for allocation.

concatMap :: Vec v a => (a -> v a) -> v a -> v a

Map a function over a vector and concatenate the results

any :: Vec v a => (a -> Bool) -> v a -> Bool

O(n) Applied to a predicate and a PrimVector, any determines if any element of the vector satisfies the predicate.

all :: Vec v a => (a -> Bool) -> v a -> Bool

O(n) Applied to a predicate and a vector, all determines if all elements of the vector satisfy the predicate.

maximum :: (Vec v a, Ord a) => v a -> a

O(n) maximum returns the maximum value from a vector It's defined with foldl1', an exception will be thrown in the case of an empty PrimVector.

minimum :: (Vec v a, Ord a) => v a -> a

O(n) minimum returns the minimum value from a vector It's defined with foldl1', an exception will be thrown in the case of an empty PrimVector.

Building vector

Scans

scanl :: forall v u a b. (Vec v a, Vec u b) => (b -> a -> b) -> b -> v a -> u b

scanl is similar to foldl, but returns a list of successive reduced values from the left.

scanl f z [x1, x2, ...] == [z, z `f` x1, (z `f` x1) `f` x2, ...]

Note that

last (scanl f z xs) == foldl f z xs.

scanl1 :: forall v a. Vec v a => (a -> a -> a) -> v a -> v a

scanl1 is a variant of scanl that has no starting value argument.

scanl1 f [x1, x2, ...] == [x1, x1 `f` x2, ...]

scanr :: forall v u a b. (Vec v a, Vec u b) => (a -> b -> b) -> b -> v a -> u b

scanr is the right-to-left dual of scanl.

scanr1 :: forall v a. Vec v a => (a -> a -> a) -> v a -> v a

scanr1 is a variant of scanr that has no starting value argument.

Accumulating maps

mapAccumL :: forall u v a b c. (Vec u b, Vec v c) => (a -> b -> (a, c)) -> a -> u b -> (a, v c)

The mapAccumL function behaves like a combination of map and foldl; it applies a function to each element of a vector, passing an accumulating parameter from left to right, and returning a final value of this accumulator together with the new list.

mapAccumR :: forall u v a b c. (Vec u b, Vec v c) => (a -> b -> (a, c)) -> a -> u b -> (a, v c)

The mapAccumR function behaves like a combination of map and foldr; it applies a function to each element of a vector, passing an accumulating parameter from right to left, and returning a final value of this accumulator together with the new vector.

Generating and unfolding vector

replicate :: Vec v a => Int -> a -> v a

unfoldr :: Vec u b => (a -> Maybe (b, a)) -> a -> u b

O(n), where n is the length of the result. The unfoldr function is analogous to the List 'unfoldr'. unfoldr builds a vector from a seed value. The function takes the element and returns Nothing if it is done producing the vector or returns Just (a,b), in which case, a is the next byte in the string, and b is the seed value for further production.

Examples:

   unfoldr (\x -> if x <= 5 then Just (x, x + 1) else Nothing) 0
== pack [0, 1, 2, 3, 4, 5]

unfoldrN :: forall v a b. Vec v b => Int -> (a -> Maybe (b, a)) -> a -> (v b, Maybe a)

O(n) Like unfoldr, unfoldrN builds a vector from a seed value. However, the length of the result is limited by the first argument to unfoldrN. This function is more efficient than unfoldr when the maximum length of the result is known.

The following equation relates unfoldrN and unfoldr:

fst (unfoldrN n f s) == take n (unfoldr f s)

Substrings

take :: Vec v a => Int -> v a -> v a

O(1) take n, applied to a vector xs, returns the prefix of xs of length n, or xs itself if n > length xs.

drop :: Vec v a => Int -> v a -> v a

O(1) drop n xs returns the suffix of xs after the first n elements, or [] if n > length xs.

slice :: Vec v a => Int -> Int -> v a -> v a

O(1) Extract a sub-range vector with give start index and length.

This function is a total function just like 'take/drop', e.g.

slice 1 3 "hello"   == "ell"
slice -1 -1 "hello" == ""
slice 2 10 "hello"  == "llo"

splitAt :: Vec v a => Int -> v a -> (v a, v a)

O(1) splitAt n xs is equivalent to (take n xs, drop n xs).

Searching vectors

Searching by equality

elem :: Vec v a => a -> v a -> Bool

notElem :: Vec v a => a -> v a -> Bool

elemIndex :: (Eq a, Vec v a) => a -> v a -> Maybe Int

Searching with a predicate

find :: Vec v a => (a -> Bool) -> v a -> Maybe a

filter :: a

partition :: a

Misc

data IPair a

Index pair type for writing high performance loops

Constructors

IPair !Int a

defaultInitSize :: Int

@defaultInitSize = 30

chunkOverhead :: Int

The memory management overhead. Currently this is tuned for GHC only.

defaultChunkSize :: Int

The chunk size used for I/O. Currently set to 32k, less the memory management overhead

smallChunkSize :: Int

The recommended chunk size. Currently set to 4k, less the memory management overhead