Std.Data.Array

Description

Unified unboxed and boxed array operations using functional dependencies.

All operations are NOT bound checked, if you need checked operations please use Std.Data.Array.Checked. It exports exactly same APIs so that you can switch between without pain.

Arr typeclass

class Arr marr arr a | arr -> marr, marr -> arr where

A typeclass to unify box & unboxed, mutable & immutable array operations.

Most of these functions simply wrap their primitive counterpart, if there's no primitive ones, we polyfilled using other operations to get the same semantics.

One exception is that shrinkMutableArr only perform closure resizing on PrimArray because current RTS support only that, shrinkMutableArr will do nothing on other array type.

It's reasonable to trust GHC with specializing & inlining these polymorphric functions. They are used across this package and perform identical to their monomophric counterpart.

Minimal complete definition

newArr, newArrWith, readArr, writeArr, setArr, indexArr, indexArrM, freezeArr, thawArr, unsafeFreezeArr, unsafeThawArr, copyArr, copyMutableArr, moveArr, cloneArr, cloneMutableArr, resizeMutableArr, shrinkMutableArr, sameMutableArr, sizeofArr, sizeofMutableArr, sameArr

Methods

newArr :: (PrimMonad m, PrimState m ~ s) => Int -> m (marr s a)

Make a new array with given size.

For boxed array, all elements are uninitialized which shall not be accessed. For primitive array, elements are just random garbage.

newArrWith :: (PrimMonad m, PrimState m ~ s) => Int -> a -> m (marr s a)

Make a new array and fill it with an initial value.

readArr :: (PrimMonad m, PrimState m ~ s) => marr s a -> Int -> m a

Index mutable array in a primitive monad.

writeArr :: (PrimMonad m, PrimState m ~ s) => marr s a -> Int -> a -> m ()

Write mutable array in a primitive monad.

setArr :: (PrimMonad m, PrimState m ~ s) => marr s a -> Int -> Int -> a -> m ()

Fill mutable array with a given value.

indexArr :: arr a -> Int -> a

Index immutable array, which is a pure operation,

indexArrM :: Monad m => arr a -> Int -> m a

Index immutable array in a primitive monad, this helps in situations that you want your indexing result is not a thunk referencing whole array.

freezeArr :: (PrimMonad m, PrimState m ~ s) => marr s a -> Int -> Int -> m (arr a)

Safely freeze mutable array by make a immutable copy of its slice.

thawArr :: (PrimMonad m, PrimState m ~ s) => arr a -> Int -> Int -> m (marr s a)

Safely thaw immutable array by make a mutable copy of its slice.

unsafeFreezeArr :: (PrimMonad m, PrimState m ~ s) => marr s a -> m (arr a)

In place freeze a mutable array, the original mutable array can not be used anymore.

unsafeThawArr :: (PrimMonad m, PrimState m ~ s) => arr a -> m (marr s a)

In place thaw a immutable array, the original immutable array can not be used anymore.

copyArr :: (PrimMonad m, PrimState m ~ s) => marr s a -> Int -> arr a -> Int -> Int -> m ()

Copy a slice of immutable array to mutable array at given offset.

copyMutableArr :: (PrimMonad m, PrimState m ~ s) => marr s a -> Int -> marr s a -> Int -> Int -> m ()

Copy a slice of mutable array to mutable array at given offset. The two mutable arrays shall no be the same one.

moveArr :: (PrimMonad m, PrimState m ~ s) => marr s a -> Int -> marr s a -> Int -> Int -> m ()

Copy a slice of mutable array to mutable array at given offset. The two mutable arrays may be the same one.

cloneArr :: arr a -> Int -> Int -> arr a

Create immutable copy.

cloneMutableArr :: (PrimMonad m, PrimState m ~ s) => marr s a -> Int -> Int -> m (marr s a)

Create mutable copy.

resizeMutableArr :: (PrimMonad m, PrimState m ~ s) => marr s a -> Int -> m (marr s a)

Resize mutable array to given size.

shrinkMutableArr :: (PrimMonad m, PrimState m ~ s) => marr s a -> Int -> m ()

Shrink mutable array to given size. This operation only works on primitive arrays. For boxed array, this is a no-op, e.g. sizeOfMutableArr will not change.

sameMutableArr :: marr s a -> marr s a -> Bool

Is two mutable array are reference equal.

sizeofArr :: arr a -> Int

Size of immutable array.

sizeofMutableArr :: (PrimMonad m, PrimState m ~ s) => marr s a -> m Int

Size of mutable array.

sameArr :: arr a -> arr a -> Bool

Is two immutable array are referencing the same one.

Note that sameArr 's result may change depending on compiler's optimizations, for example let arr = runST ... in arr sameArr arr may return false if compiler decides to inline it.

See https://ghc.haskell.org/trac/ghc/ticket/13908 for more background.

data RealWorld :: *

Boxed array type

data Array a :: * -> *

Constructors

Array

Fields

array# :: Array# a

data MutableArray s a :: * -> * -> *

Constructors

MutableArray

Fields

marray# :: MutableArray# s a

data SmallArray a :: * -> *

Constructors

SmallArray (SmallArray# a)

data SmallMutableArray s a :: * -> * -> *

Constructors

SmallMutableArray (SmallMutableArray# s a)

uninitialized :: a

Bottom value (throw (UndefinedElement "Data.Array.uninitialized")) for initialize new boxed array(Array, SmallArray..).

Primitive array type

data PrimArray a :: * -> *

Constructors

PrimArray ByteArray#

data MutablePrimArray s a :: * -> * -> *

Constructors

MutablePrimArray (MutableByteArray# s)

class Prim a where

Minimal complete definition

sizeOf#, alignment#, indexByteArray#, readByteArray#, writeByteArray#, setByteArray#, indexOffAddr#, readOffAddr#, writeOffAddr#, setOffAddr#

Methods

sizeOf# :: a -> Int#

alignment# :: a -> Int#

indexByteArray# :: ByteArray# -> Int# -> a

readByteArray# :: MutableByteArray# s -> Int# -> State# s -> (#VoidRep, PtrRepLifted, State# s, a#)

writeByteArray# :: MutableByteArray# s -> Int# -> a -> State# s -> State# s

setByteArray# :: MutableByteArray# s -> Int# -> Int# -> a -> State# s -> State# s

indexOffAddr# :: Addr# -> Int# -> a

readOffAddr# :: Addr# -> Int# -> State# s -> (#VoidRep, PtrRepLifted, State# s, a#)

writeOffAddr# :: Addr# -> Int# -> a -> State# s -> State# s

setOffAddr# :: Addr# -> Int# -> Int# -> a -> State# s -> State# s

newPinnedPrimArray :: forall m a. (PrimMonad m, Prim a) => Int -> m (MutablePrimArray (PrimState m) a)

Create a pinned byte array of the specified size, The garbage collector is guaranteed not to move it.

newAlignedPinnedPrimArray :: forall m a. (PrimMonad m, Prim a) => Int -> m (MutablePrimArray (PrimState m) a)

Create a pinned primitive array of the specified size and respect given primitive type's alignment. The garbage collector is guaranteed not to move it.

copyPrimArrayToPtr :: (PrimMonad m, Prim a) => Ptr a -> PrimArray a -> Int -> Int -> m ()

copyMutablePrimArrayToPtr :: (PrimMonad m, Prim a) => Ptr a -> MutablePrimArray (PrimState m) a -> Int -> Int -> m ()

copyMutablePrimArrayFromPtr

Arguments

:: (PrimMonad m, Prim a)

=> MutablePrimArray (PrimState m) a
destination array

-> Int
offset into destination array

-> Ptr a
source pointer

-> Int
number of prims to copy

-> m ()

Copy a slice of an mutable primitive array from an address. The offset and length are given in elements of type a.

primArrayContents :: PrimArray a -> Ptr a

Yield a pointer to the array's data.

This operation is only safe on pinned primitive arrays allocated by newPinnedPrimArray or newAlignedPinnedPrimArray, and you have to make sure the PrimArray can outlive the Ptr.

mutablePrimArrayContents :: MutablePrimArray s a -> Ptr a

Yield a pointer to the array's data.

This operation is only safe on pinned primitive arrays allocated by newPinnedPrimArray or newAlignedPinnedPrimArray. and you have to make sure the PrimArray can outlive the Ptr.

withPrimArrayContents :: PrimArray a -> (Ptr a -> IO b) -> IO b

Yield a pointer to the array's data and do computation with it.

This operation is only safe on pinned primitive arrays allocated by newPinnedPrimArray or newAlignedPinnedPrimArray.

withMutablePrimArrayContents :: MutablePrimArray RealWorld a -> (Ptr a -> IO b) -> IO b

Yield a pointer to the array's data and do computation with it.

This operation is only safe on pinned primitive arrays allocated by newPinnedPrimArray or newAlignedPinnedPrimArray.

isPrimArrayPinned :: PrimArray a -> Bool

Check if a primitive array is pinned.

isMutablePrimArrayPinned :: MutablePrimArray s a -> Bool

Check if a mutable primitive array is pinned.

Unlifted array type

data UnliftedArray e :: * -> *

Constructors

UnliftedArray ArrayArray#

data MutableUnliftedArray s e :: * -> * -> *

Constructors

MutableUnliftedArray (MutableArrayArray# s)

class PrimUnlifted a where

Minimal complete definition

toArrayArray#, fromArrayArray#

Methods

toArrayArray# :: a -> ArrayArray#

fromArrayArray# :: ArrayArray# -> a

The ArrayException type

data ArrayException :: *

Constructors

IndexOutOfBounds String

UndefinedElement String