Std.Data.Binary

Description

Efficiently serialization into builder using binary encoding. Default instances use "Intel convention", i.e. little endian encoding, encoded data should be portable across machine endianness, word size, or compiler version within one major version. For example, data encoded using the Binary class could be written on any machine, and read back on any another using stdio packages with the same major version.

Documentation

class Binary a where

Minimal complete definition

binary

Methods

binary :: a -> Builder

class Binary a => PrimPut a where

Minimal complete definition

boundedSize, boundedWrite

Methods

boundedSize :: a -> Int

boundedWrite :: a -> MutablePrimArray RealWorld Word8 -> Int -> IO Int

primBinary :: PrimPut a => a -> Builder

newtype BE a

A newtype wrapper for big endian's instances.

Constructors

BE a

class BinaryParse a where

Minimal complete definition

binaryParse

Methods

binaryParse :: Parser a