Std.Data.Builder

Description

A Builder records a buffer writing function, which can be mappend in O(1) via composition. This module provides many functions to turn basic data types into Builders, which can used to build strict Bytes or list of Bytes chunks.

Builder type

data Builder a

Builder is a monad to help compose BuilderStep. With next BuilderStep continuation, we can do interesting things like perform some action, or interleave the build process.

Notes on IsString instance: Builder '()''s IsString instance use stringModifiedUTF8, which is different from stringUTF8 in that it DOES NOT PROVIDE UTF8 GUARANTEES! :

NUL will be written as xC0 x80.

xD800 ~ xDFFF will be encoded in three bytes as normal UTF-8 codepoints.

Because this is also how ghc compile string literal into binaries.

append :: Builder a -> Builder b -> Builder b

Running builders

buildBytes :: Builder a -> Bytes

shortcut to buildBytesWith defaultInitSize.

buildBytesWith :: Int -> Builder a -> Bytes

run Builder with DoubleBuffer strategy, which is suitable for building short bytes.

buildBytesList :: Builder a -> [Bytes]

shortcut to buildBytesListWith defaultChunkSize.

buildBytesListWith :: Int -> Int -> Builder a -> [Bytes]

run Builder with InsertChunk strategy, which is suitable for building lazy bytes chunks.

buildAndRun :: (Bytes -> IO ()) -> Builder a -> IO ()

shortcut to buildAndRunWith defaultChunkSize.

buildAndRunWith :: Int -> (Bytes -> IO ()) -> Builder a -> IO ()

run Builder with OneShotAction strategy, which is suitable for doing effects while building.

Basic buiders

bytes :: Bytes -> Builder ()

Write a Bytes.

ensureN :: Int -> Builder ()

Ensure that there are at least n many elements available.

atMost

Arguments

:: Int
size bound

-> (forall s. MutablePrimArray s Word8 -> Int -> ST s Int)
the writer which return a new offset for next write

-> Builder ()

writeN

Arguments

:: Int
size bound

-> (forall s. MutablePrimArray s Word8 -> Int -> ST s ())
the writer which return a new offset for next write

-> Builder ()

Pritimive builders

encodePrim :: forall a. UnalignedAccess a => a -> Builder ()

write primitive types in host byte order.

encodePrimLE :: forall a. UnalignedAccess (LE a) => a -> Builder ()

write primitive types with little endianess.

encodePrimBE :: forall a. UnalignedAccess (BE a) => a -> Builder ()

write primitive types with big endianess.

More builders

stringUTF8 :: String -> Builder ()

Turn String into Builder with UTF8 encoding

Illegal codepoints will be written as replacementChars.

Note, if you're trying to write string literals builders, and you know it doen't contain '\NUL' or surrgate codepoints, then you can open OverloadedStrings and use Builder's IsString instance, it can save an extra UTF-8 validation.

This function will be rewritten into a memcpy if possible, (running a fast UTF-8 validation at runtime first).

charUTF8 :: Char -> Builder ()

Turn Char into Builder with UTF8 encoding

Illegal codepoints will be written as replacementChars.

string7 :: String -> Builder ()

Turn String into Builder with ASCII7 encoding

Codepoints beyond '\x7F' will be chopped.

char7 :: Char -> Builder ()

Turn Char into Builder with ASCII7 encoding

Codepoints beyond '\x7F' will be chopped.

string8 :: String -> Builder ()

Turn String into Builder with ASCII8 encoding

Codepoints beyond '\xFF' will be chopped. Note, this encoding is NOT compatible with UTF8 encoding, i.e. bytes written by this builder may not be legal UTF8 encoding bytes.

char8 :: Char -> Builder ()

Turn Char into Builder with ASCII8 encoding

Codepoints beyond '\xFF' will be chopped. Note, this encoding is NOT compatible with UTF8 encoding, i.e. bytes written by this builder may not be legal UTF8 encoding bytes.

text :: Text -> Builder ()

Write UTF8 encoded Text using Builder.

Note, if you're trying to write string literals builders, please open OverloadedStrings and use Builders IsString instance, it will be rewritten into a memcpy.

Numeric builders

Integral type formatting

data IFormat

Integral formatting options.

Constructors

IFormat

Fields

width :: Int

total width, only effective with padding options

padding :: Padding

padding options

postiveSign :: Bool

show + when the number is positive

defaultIFormat :: IFormat

defaultIFormat = IFormat 0 NoPadding False Decimal

data Padding

Constructors

NoPadding

ZeroPadding

LeftSpacePadding

RightSpacePadding

int :: (Integral a, Bounded a) => a -> Builder ()

int = intWith defaultIFormat

intWith :: (Integral a, Bounded a) => IFormat -> a -> Builder ()

Format a Bounded Integral type like Int or Word16 into decimal ASCII digits.

integer :: Integer -> Builder ()

Format a Integer into decimal ASCII digits.

Fixded size hexidecimal formatting

hex :: forall a. (FiniteBits a, Integral a) => a -> Builder ()

Format a FiniteBits Integral type into hex nibbles.

heX :: forall a. (FiniteBits a, Integral a) => a -> Builder ()

The UPPERCASED version of hex.

IEEE float formating

data FFormat

Control the rendering of floating point numbers.

Constructors

Exponent
Scientific notation (e.g. 2.3e123).

Fixed
Standard decimal notation.

Generic
Use decimal notation for values between 0.1 and 9,999,999, and scientific notation otherwise.

double :: Double -> Builder ()

Decimal encoding of an IEEE Double.

Using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

doubleWith

Arguments

:: FFormat

-> Maybe Int
Number of decimal places to render.

-> Double

-> Builder ()

Format double-precision float using drisu3 with dragon4 fallback.

float :: Float -> Builder ()

Decimal encoding of an IEEE Float.

Using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

floatWith

Arguments

:: FFormat

-> Maybe Int
Number of decimal places to render.

-> Float

-> Builder ()

Format single-precision float using drisu3 with dragon4 fallback.

scientific :: Scientific -> Builder ()

A Builder which renders a scientific number to full precision, using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

scientificWith

Arguments

:: FFormat

-> Maybe Int
Number of decimal places to render.

-> Scientific

-> Builder ()

Like scientific but provides rendering options.