Std.Data.Builder.Numeric

Description

Textual numeric builders.

Integral type formatting

data IFormat

Integral formatting options.

Constructors

IFormat

Fields

width :: Int

total width, only effective with padding options

padding :: Padding

padding options

postiveSign :: Bool

show + when the number is positive

defaultIFormat :: IFormat

defaultIFormat = IFormat 0 NoPadding False Decimal

data Padding

Constructors

NoPadding

ZeroPadding

LeftSpacePadding

RightSpacePadding

int :: (Integral a, Bounded a) => a -> Builder ()

int = intWith defaultIFormat

intWith :: (Integral a, Bounded a) => IFormat -> a -> Builder ()

Format a Bounded Integral type like Int or Word16 into decimal ASCII digits.

integer :: Integer -> Builder ()

Format a Integer into decimal ASCII digits.

Fixded size hexidecimal formatting

hex :: forall a. (FiniteBits a, Integral a) => a -> Builder ()

Format a FiniteBits Integral type into hex nibbles.

heX :: forall a. (FiniteBits a, Integral a) => a -> Builder ()

The UPPERCASED version of hex.

IEEE float formating

data FFormat

Control the rendering of floating point numbers.

Constructors

Exponent
Scientific notation (e.g. 2.3e123).

Fixed
Standard decimal notation.

Generic
Use decimal notation for values between 0.1 and 9,999,999, and scientific notation otherwise.

double :: Double -> Builder ()

Decimal encoding of an IEEE Double.

Using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

doubleWith

Arguments

:: FFormat

-> Maybe Int
Number of decimal places to render.

-> Double

-> Builder ()

Format double-precision float using drisu3 with dragon4 fallback.

float :: Float -> Builder ()

Decimal encoding of an IEEE Float.

Using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

floatWith

Arguments

:: FFormat

-> Maybe Int
Number of decimal places to render.

-> Float

-> Builder ()

Format single-precision float using drisu3 with dragon4 fallback.

scientific :: Scientific -> Builder ()

A Builder which renders a scientific number to full precision, using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

scientificWith

Arguments

:: FFormat

-> Maybe Int
Number of decimal places to render.

-> Scientific

-> Builder ()

Like scientific but provides rendering options.

Misc

grisu3 :: Double -> Maybe ([Int], Int)

Decimal encoding of a Double.

grisu3_sp :: Float -> Maybe ([Int], Int)

Decimal encoding of a Float.

i2wDec :: Integral a => a -> Word8

Decimal digit to ASCII digit.

i2wHex :: Integral a => a -> Word8

Hexadecimal digit to ASCII char.

i2wHeX :: Integral a => a -> Word8

Hexadecimal digit to UPPERCASED ASCII char.

countDigits :: Integral a => a -> Int

Count how many decimal digits an integer has.