Std.Data.LEON

Description

LEON (little endian object notation) is an efficiently serialization using simple binary encoding. As suggested by its name, default instances use little endian encoding, i.e. "Intel convention", encoded data should be portable across machine endianness, word size, or compiler version within one major version. For example, data encoded using the LEON class could be written on any machine, and read back on any another using stdio packages with the same major version.

Documentation

class GLEONEncode f where

Methods

gencode :: f t -> Builder ()

class GLEONDecode f where

Methods

gdecode :: Parser (f t)

class LEON a where

LEON, Little Endian Object Notation.

Minimal complete definition

Nothing

Methods

encode :: a -> Builder ()

decode :: Parser a

encode :: (Generic a, GLEONEncode (Rep a)) => a -> Builder ()

decode :: (Generic a, GLEONDecode (Rep a)) => Parser a

sizeError :: Show size => String -> size -> error

checkGetSum :: (Ord word, Num word, Bits word, GSumDecode f) => word -> word -> Parser (f a)

class GSumDecode f where

Methods

decodeSum :: (Ord word, Num word, Bits word) => word -> word -> Parser (f a)

class GSumEncode f where

Methods

encodeSum :: (Num w, Bits w, LEON w) => w -> w -> f a -> Builder ()

class SumSize f where

Methods

sumSize :: Tagged f Word64

newtype Tagged (s :: Type -> Type) b

Constructors

Tagged

Fields

unTagged :: b