Std.Data.PrimArray.BitTwiddle

Description

This module implement some bit twiddling with ghc primitives.

We currently didn't use all functions from this module though: the performance is not catching up c version yet. But this module and relevant benchmarks are kept in hope that once we have fully SIMD support in GHC, we might optimize these functions further to compete with c.

Reference:

https://graphics.stanford.edu/~seander/bithacks.html

https://jameshfisher.github.io/2017/01/24/bitwise-check-for-zero-byte.html

Documentation

isOffsetAligned# :: Int# -> Bool

mkMask# :: Word# -> Word#

nullByteMagic# :: Word# -> Word#

memchr :: PrimArray Word8 -> Word8 -> Int -> Int -> Int

Search a word8 in array.

Currently this function is ~4 times slow than c version, so we didn't use it.

memchr# :: ByteArray# -> Word# -> Int# -> Int# -> Int#

The unboxed version of memchr

memchrReverse :: PrimArray Word8 -> Word8 -> Int -> Int -> Int

Search a word8 array in reverse order.

This function is used in elemIndexEnd, since there's no c equivalent.

memchrReverse# :: ByteArray# -> Word# -> Int# -> Int# -> Int#

The unboxed version of memchrReverse

c_memchr :: ByteArray# -> Int -> Word8 -> Int -> Int