Std.Data.PrimArray.Cast

Description

This module is borrowed from basement's Cast module with conditional instances removed. The purpose of Cast is to provide primitive types which share the same byte size, so that arrays and vectors parameterized by them can be safely coerced without breaking the index bounds. You can also use it to directly cast primitives just like reinterpret_cast. A Coercible based instance is also provide for convenience.

Documentation

class Cast source destination where

Cast between primitive types of the same size.

Methods

cast :: source -> destination