Std.Data.PrimSTRef.Base

Description

Internal module for PrimSTRef and PrimIORef.

Unboxed ST references

newtype PrimSTRef s a

A mutable variable in the ST monad which can hold an instance of Prim.

Constructors

PrimSTRef (MutableByteArray s)

newPrimSTRef :: Prim a => a -> ST s (PrimSTRef s a)

Build a new PrimSTRef

readPrimSTRef :: Prim a => PrimSTRef s a -> ST s a

Read the value of an PrimSTRef

writePrimSTRef :: Prim a => PrimSTRef s a -> a -> ST s ()

Write a new value into an PrimSTRef

modifyPrimSTRef :: Prim a => PrimSTRef s a -> (a -> a) -> ST s ()

Mutate the contents of an PrimSTRef.

Unboxed reference is always strict on the value it hold.