Std.Data.Text.Extra

Description

Various combinators works on Texts.

Slice manipulation

cons :: Char -> Text -> Text

snoc :: Text -> Char -> Text

uncons :: Text -> Maybe (Char, Text)

unsnoc :: Text -> Maybe (Text, Char)

headMaybe :: Text -> Maybe Char

tailMayEmpty :: Text -> Text

lastMaybe :: Text -> Maybe Char

initMayEmpty :: Text -> Text

inits :: Text -> [Text]

tails :: Text -> [Text]

take :: Int -> Text -> Text

drop :: Int -> Text -> Text

takeLast :: Int -> Text -> Text

dropLast :: Int -> Text -> Text

slice :: Int -> Int -> Text -> Text

O(1) Extract a sub-range text with give start index and length.

This function is a total function just like 'take/drop', index/length exceeds range will be ingored, e.g.

slice 1 3 "hello"   == "ell"
slice -1 -1 "hello" == ""
slice -2 2 "hello"  == ""
slice 2 10 "hello"  == "llo"

This holds for all x y: slice x y vs == drop x . take (x+y) vs

splitAt :: Int -> Text -> (Text, Text)

takeWhile :: (Char -> Bool) -> Text -> Text

O(n) Applied to a predicate p and a text t, returns the longest prefix (possibly empty) of t of elements that satisfy p.

takeLastWhile :: (Char -> Bool) -> Text -> Text

O(n) Applied to a predicate p and a text t, returns the longest suffix (possibly empty) of t of elements that satisfy p.

dropWhile :: (Char -> Bool) -> Text -> Text

dropLastWhile :: (Char -> Bool) -> Text -> Text

dropAround :: (Char -> Bool) -> Text -> Text

break :: (Char -> Bool) -> Text -> (Text, Text)

span :: (Char -> Bool) -> Text -> (Text, Text)

breakEnd :: (Char -> Bool) -> Text -> (Text, Text)

spanEnd :: (Char -> Bool) -> Text -> (Text, Text)

breakOn :: Text -> Text -> (Text, Text)

breakOnAll :: Text -> Text -> [(Text, Text)]

breakOnAllOverlapping :: Text -> Text -> [(Text, Text)]

group :: Text -> [Text]

groupBy :: (Char -> Char -> Bool) -> Text -> [Text]

stripPrefix :: Text -> Text -> Maybe Text

O(n) The stripPrefix function takes two texts and returns Just the remainder of the second iff the first is its prefix, and otherwise Nothing.

stripSuffix :: Text -> Text -> Maybe Text

split :: Char -> Text -> [Text]

splitWith :: (Char -> Bool) -> Text -> [Text]

O(n) Splits a text into components delimited by separators, where the predicate returns True for a separator char. The resulting components do not contain the separators. Two adjacent separators result in an empty component in the output. eg.

splitWith (=='a') "aabbaca" == ["","","bb","c",""]
splitWith (=='a') []        == [""]

splitOn :: Text -> Text -> [Text]

isPrefixOf :: Text -> Text -> Bool

isSuffixOf :: Text -> Text -> Bool

isInfixOf :: Text -> Text -> Bool

commonPrefix :: Text -> Text -> (Text, Text, Text)

O(n) Find the longest non-empty common prefix of two strings and return it, along with the suffixes of each string at which they no longer match. e.g.

>>> commonPrefix "foobar" "fooquux"
("foo","bar","quux")

>>> commonPrefix "veeble" "fetzer"
("","veeble","fetzer")

words :: Text -> [Text]

O(n) Breaks a Bytes up into a list of words, delimited by unicode space.

lines :: Text -> [Text]

unwords :: [Text] -> Text

unlines :: [Text] -> Text

padLeft :: Int -> Char -> Text -> Text

padRight :: Int -> Char -> Text -> Text

Transform

reverse :: Text -> Text

O(n) Reverse the characters of a string.

intersperse :: Char -> Text -> Text

O(n) The intersperse function takes a character and places it between the characters of a Text. Performs replacement on invalid scalar values.

intercalate :: Text -> [Text] -> Text

O(n) The intercalate function takes a Text and a list of Texts and concatenates the list after interspersing the first argument between each element of the list.

intercalateElem :: Char -> [Text] -> Text

transpose :: [Text] -> [Text]

The transpose function transposes the rows and columns of its text argument.