Std.Data.Text.Search

Description

Searching by equality

elem :: Char -> Text -> Bool

O(n) elem test if given char is in given text.

notElem :: Char -> Text -> Bool

O(n) 'not . elem'

element-wise search

findIndices :: (Char -> Bool) -> Text -> [Int]

find

Arguments

:: (Char -> Bool)

-> Text

-> (Int, Int, Maybe Char)
(char index, byte index, matching char)

O(n) find the first char matching the predicate in a text from left to right, if there isn't one, return the index point to the end of the byte slice.

findLast :: (Char -> Bool) -> Text -> (Int, Int, Maybe Char)

O(n) find the first char matching the predicate in a text from right to left, if there isn't one, return the index point to the end of the byte slice.

findIndexOrEnd :: (Char -> Bool) -> Text -> Int

findLastIndexOrStart :: (Char -> Bool) -> Text -> Int

filter :: (Char -> Bool) -> Text -> Text

partition :: (Char -> Bool) -> Text -> (Text, Text)

O(n) The partition function takes a predicate, a text, returns a pair of text with codepoints which do and do not satisfy the predicate, respectively; i.e.,

partition p txt == (filter p txt, filter (not . p) txt)