Std.Data.TextBuilder

Description

UTF8 compatible textual builders.

Textual Builder

newtype TextBuilder a

Buidlers which guarantee UTF-8 encoding, thus can be used to build text directly.

Notes on IsString instance: It's recommended to use IsString instance instead of stringUTF8 or string7 since there's a rewrite rule to turn encoding loop into a memcpy, which is much faster.

Different from 'Builder ()', 'TextBuilder ()''s IsString instance will gives desired UTF8 guarantees:

"NUL" will be written directly as x00.

xD800 ~ xDFFF will be replaced by replacement char.

Constructors

TextBuilder

Fields

toBuilder :: Builder a

buildText :: TextBuilder a -> Text

Basic UTF8 builders

stringUTF8 :: String -> TextBuilder ()

Turn String into TextBuilder with UTF8 encoding

Illegal codepoints will be written as replacementChars.

Note, if you're trying to write string literals builders, please open OverloadedStrings and use TextBuilder's IsString instance, it will be rewritten into a memcpy, instead of encoding Chars in a loop like what stringUTF8 do.

charUTF8 :: Char -> TextBuilder ()

Turn Char into TextBuilder with UTF8 encoding

Illegal codepoints will be written as replacementChars.

string7 :: String -> TextBuilder ()

Turn String into TextBuilder with ASCII7 encoding

Codepoints beyond '\x7F' will be chopped.

char7 :: Char -> TextBuilder ()

Turn Char into TextBuilder with ASCII7 encoding

Codepoints beyond '\x7F' will be chopped.

text :: Text -> TextBuilder ()

Write UTF8 encoded Text using Builder.

Note, if you're trying to write string literals builders, please open OverloadedStrings and use Builders IsString instance, it will be rewritten into a memcpy.

Numeric builders

Integral type formatting

data IFormat

Integral formatting options.

Constructors

IFormat

Fields

width :: Int

total width, only effective with padding options

padding :: Padding

padding options

postiveSign :: Bool

show + when the number is positive

defaultIFormat :: IFormat

defaultIFormat = IFormat 0 NoPadding False Decimal

data Padding

Constructors

NoPadding

ZeroPadding

LeftSpacePadding

RightSpacePadding

int :: (Integral a, Bounded a) => a -> TextBuilder ()

int = intWith defaultIFormat

intWith :: (Integral a, Bounded a) => IFormat -> a -> TextBuilder ()

Format a Bounded Integral type like Int or Word16 into decimal ascii digits.

integer :: Integer -> TextBuilder ()

Format a Integer into decimal ascii digits.

Fixded size hexidecimal formatting

hex :: (FiniteBits a, Integral a) => a -> TextBuilder ()

Format a FiniteBits Integral type into hex nibbles.

heX :: (FiniteBits a, Integral a) => a -> TextBuilder ()

The UPPERCASED version of hex.

IEEE float formating

data FFormat

Control the rendering of floating point numbers.

Constructors

Exponent
Scientific notation (e.g. 2.3e123).

Fixed
Standard decimal notation.

Generic
Use decimal notation for values between 0.1 and 9,999,999, and scientific notation otherwise.

double :: Double -> TextBuilder ()

Decimal encoding of an IEEE Double.

Using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

doubleWith

Arguments

:: FFormat

-> Maybe Int
Number of decimal places to render.

-> Double

-> TextBuilder ()

Format double-precision float using drisu3 with dragon4 fallback.

float :: Float -> TextBuilder ()

Decimal encoding of an IEEE Float.

Using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

floatWith

Arguments

:: FFormat

-> Maybe Int
Number of decimal places to render.

-> Float

-> TextBuilder ()

Format single-precision float using drisu3 with dragon4 fallback.

scientific :: Scientific -> TextBuilder ()

A Builder which renders a scientific number to full precision, using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

scientificWith

Arguments

:: FFormat

-> Maybe Int
Number of decimal places to render.

-> Scientific

-> TextBuilder ()

Like scientific but provides rendering options.