Std.Data.Vector.Base

Description

This module provides unified vector interface. Conceptually a vector is simply a slice of an array, for example this is the definition of boxed vector:

data Vector a = Vector !(SmallArray a)   !Int    !Int
                     -- payload           offset  length

The Vec class unified different type of vectors, and this module provide operation over Vec instances, with all the internal structures. Be careful on modifying internal slices, otherwise segmentation fault await.

The Vec typeclass

class Arr (MArray v) (IArray v) a => Vec v a where

Typeclass for box and unboxed vectors, which are created by slicing arrays.

Associated Types

type MArray v = (marr :: * -> * -> *) | marr -> v

Vector's mutable array type

type IArray v = (iarr :: * -> *) | iarr -> v

Vector's immutable array type

Methods

toArr :: v a -> (IArray v a, Int, Int)

Get underline array and slice range(offset and length).

fromArr :: IArray v a -> Int -> Int -> v a

Create a vector by slicing an array(with offset and length).

pattern Vec :: Vec v a => IArray v a -> Int -> Int -> v a

A pattern synonyms for matching the underline array, offset and length.

This is a bidirectional pattern synonyms, but very unsafe if not use properly. Make sure your slice is within array's bounds!

indexMaybe :: (Vec v a, HasCallStack) => v a -> Int -> Maybe a

O(1) Index array element.

Return Nothing if index is out of bounds.

Boxed and unboxed vector type

data Vector a

Boxed vector

Constructors

Vector !(SmallArray a) !Int !Int

data PrimVector a

Primitive vector

Constructors

PrimVector !(PrimArray a) !Int !Int

Word8 vector

type Bytes = PrimVector Word8

Bytes is just primitive word8 vectors.

w2c :: Word8 -> Char

Conversion between Word8 and Char. Should compile to a no-op.

c2w :: Char -> Word8

Unsafe conversion between Char and Word8. This is a no-op and silently truncates to 8 bits Chars > '\255'. It is provided as convenience for PrimVector construction.

Basic creating

create

Arguments

:: Vec v a

=> Int
length in elements of type a

-> (forall s. MArray v s a -> ST s ())
initialization function

-> v a

Create a vector with size N.

create'

Arguments

:: Vec v a

=> Int
length in elements of type a

-> (forall s. MArray v s a -> ST s (IPair (MArray v s a)))
initialization function return a result size and array the result must start from index 0

-> v a

Create a vector with a initial size N array (which may not be the final array).

creating

Arguments

:: Vec v a

=> Int

-> (forall s. MArray v s a -> ST s b)
initialization function

-> (b, v a)

Create a vector with a initial size N array, return both the vector and the monadic result during creating.

The result is not demanded strictly while the returned vector will be in normal form. It this is not desired, use return $! idiom in your initialization function.

creating'

Arguments

:: Vec v a

=> Int

-> (forall s. MArray v s a -> ST s (b, IPair (MArray v s a)))
initialization function

-> (b, v a)

Create a vector with a initial size N array (which may not be the final array), return both the vector and the monadic result during creating.

The result is not demanded strictly while the returned vector will be in normal form. It this is not desired, use return $! idiom in your initialization function.

createN

Arguments

:: (Vec v a, HasCallStack)

=> Int
length's upper bound

-> (forall s. MArray v s a -> ST s Int)
initialization function which return the actual length

-> v a

Create a vector up to a specific length.

If the initialization function return a length larger than initial size, an IndexOutOfVectorRange will be raised.

createN2 :: (Vec v a, Vec u b, HasCallStack) => Int -> Int -> (forall s. MArray v s a -> MArray u s b -> ST s (Int, Int)) -> (v a, u b)

Create two vector up to a specific length.

If the initialization function return lengths larger than initial sizes, an IndexOutOfVectorRange will be raised.

empty :: Vec v a => v a

O(1). The empty vector.

singleton :: Vec v a => a -> v a

O(1). Single element vector.

copy :: Vec v a => v a -> v a

O(n). Copy a vector from slice.

Conversion between list

pack :: Vec v a => [a] -> v a

O(n) Convert a list into a vector

Alias for packN defaultInitSize.

packN :: forall v a. Vec v a => Int -> [a] -> v a

O(n) Convert a list into a vector with an approximate size.

If the list's length is large than the size given, we simply double the buffer size and continue building.

This function is a good consumer in the sense of build/foldr fusion.

packR :: Vec v a => [a] -> v a

O(n)

Alias for packRN defaultInitSize.

packRN :: forall v a. Vec v a => Int -> [a] -> v a

O(n) packN in reverse order.

This function is a good consumer in the sense of build/foldr fusion.

unpack :: Vec v a => v a -> [a]

O(n) Convert vector to a list.

Unpacking is done lazily. i.e. we will retain reference to the array until all element are consumed.

This function is a good producer in the sense of build/foldr fusion.

unpackR :: Vec v a => v a -> [a]

O(n) Convert vector to a list in reverse order.

This function is a good producer in the sense of build/foldr fusion.

Basic interface

null :: Vec v a => v a -> Bool

O(1) Test whether a vector is empty.

length :: Vec v a => v a -> Int

O(1) The length of a vector.

append :: Vec v a => v a -> v a -> v a

O(m+n)

There's no need to guard empty vector because we guard them for you, so appending empty vectors are no-ops.

map :: forall u v a b. (Vec u a, Vec v b) => (a -> b) -> u a -> v b

Mapping between vectors (possiblely with two different vector types).

NOTE, the result vector contain thunks in lifted Vector case, use map' if that's not desired.

For PrimVector, map and map' are same, since PrimVectors never store thunks.

map' :: forall u v a b. (Vec u a, Vec v b) => (a -> b) -> u a -> v b

Mapping between vectors (possiblely with two different vector types).

This is the strict version map. Note that the Functor instance of lifted Vector is defined with map to statisfy laws, which this strict version breaks (map' id arrayContainsBottom /= arrayContainsBottom ).

imap' :: forall u v a b. (Vec u a, Vec v b) => (Int -> a -> b) -> u a -> v b

Strict mapping with index.

foldl' :: Vec v a => (b -> a -> b) -> b -> v a -> b

Strict left to right fold.

ifoldl' :: Vec v a => (b -> Int -> a -> b) -> b -> v a -> b

Strict left to right fold with index.

foldl1' :: forall v a. (Vec v a, HasCallStack) => (a -> a -> a) -> v a -> a

Strict left to right fold using first element as the initial value.

Throw EmptyVector if vector is empty.

foldl1Maybe' :: forall v a. Vec v a => (a -> a -> a) -> v a -> Maybe a

Strict left to right fold using first element as the initial value. return Nothing when vector is empty.

foldr' :: Vec v a => (a -> b -> b) -> b -> v a -> b

Strict right to left fold

ifoldr' :: Vec v a => (Int -> a -> b -> b) -> b -> v a -> b

Strict right to left fold with index

NOTE: the index is counting from 0, not backwards

foldr1' :: forall v a. (Vec v a, HasCallStack) => (a -> a -> a) -> v a -> a

Strict right to left fold using last element as the initial value.

Throw EmptyVector if vector is empty.

foldr1Maybe' :: forall v a. Vec v a => (a -> a -> a) -> v a -> Maybe a

Strict right to left fold using last element as the initial value, return Nothing when vector is empty.

Special folds

concat :: forall v a. Vec v a => [v a] -> v a

O(n) Concatenate a list of vector.

Note: concat have to force the entire list to filter out empty vector and calculate the length for allocation.

concatMap :: Vec v a => (a -> v a) -> v a -> v a

Map a function over a vector and concatenate the results

maximum :: (Vec v a, Ord a, HasCallStack) => v a -> a

O(n) maximum returns the maximum value from a vector

It's defined with foldl1', an EmptyVector exception will be thrown in the case of an empty vector.

minimum :: (Vec v a, Ord a, HasCallStack) => v a -> a

O(n) minimum returns the minimum value from a vector

An EmptyVector exception will be thrown in the case of an empty vector.

maximumMaybe :: (Vec v a, Ord a, HasCallStack) => v a -> Maybe a

O(n) maximum returns the maximum value from a vector, return Nothing in the case of an empty vector.

minimumMaybe :: (Vec v a, Ord a, HasCallStack) => v a -> Maybe a

O(n) minimum returns the minimum value from a vector, return Nothing in the case of an empty vector.

sum :: (Vec v a, Num a) => v a -> a

O(n) sum returns the sum value from a vector

count :: (Vec v a, Eq a) => a -> v a -> Int

O(n) count returns count of an element from a vector

product :: (Vec v a, Num a) => v a -> a

O(n) product returns the product value from a vector

product' :: (Vec v a, Num a, Eq a) => v a -> a

O(n) product returns the product value from a vector

This function will shortcut on zero. Note this behavior change the semantics for lifted vector: product [1,0,undefined] /= product' [1,0,undefined].

all :: Vec v a => (a -> Bool) -> v a -> Bool

O(n) Applied to a predicate and a vector, all determines if all elements of the vector satisfy the predicate.

any :: Vec v a => (a -> Bool) -> v a -> Bool

O(n) Applied to a predicate and a vector, any determines if any elements of the vector satisfy the predicate.

Building vector

Accumulating maps

mapAccumL :: forall u v a b c. (Vec u b, Vec v c) => (a -> b -> (a, c)) -> a -> u b -> (a, v c)

The mapAccumL function behaves like a combination of map and foldl; it applies a function to each element of a vector, passing an accumulating parameter from left to right, and returning a final value of this accumulator together with the new list.

Note, this function will only force the result tuple, not the elements inside, to prevent creating thunks during mapAccumL, seq your accumulator and result with the result tuple.

mapAccumR :: forall u v a b c. (Vec u b, Vec v c) => (a -> b -> (a, c)) -> a -> u b -> (a, v c)

The mapAccumR function behaves like a combination of map and foldr; it applies a function to each element of a vector, passing an accumulating parameter from right to left, and returning a final value of this accumulator together with the new vector.

The same strictness property with mapAccumL applys to mapAccumR too.

Generating and unfolding vector

replicate :: Vec v a => Int -> a -> v a

O(n) replicate n x is a vector of length n with x the value of every element.

Note: replicate will not force the element.

cycleN :: forall v a. Vec v a => Int -> v a -> v a

O(n*m) cycleN a vector n times.

unfoldr :: Vec u b => (a -> Maybe (b, a)) -> a -> u b

O(n), where n is the length of the result. The unfoldr function is analogous to the List 'unfoldr'. unfoldr builds a vector from a seed value. The function takes the element and returns Nothing if it is done producing the vector or returns Just (a,b), in which case, a is the next byte in the string, and b is the seed value for further production.

Examples:

   unfoldr (\x -> if x <= 5 then Just (x, x + 1) else Nothing) 0
== pack [0, 1, 2, 3, 4, 5]

unfoldrN :: forall v a b. Vec v b => Int -> (a -> Maybe (b, a)) -> a -> (v b, Maybe a)

O(n) Like unfoldr, unfoldrN builds a vector from a seed value. However, the length of the result is limited by the first argument to unfoldrN. This function is more efficient than unfoldr when the maximum length of the result is known.

The following equation relates unfoldrN and unfoldr:

fst (unfoldrN n f s) == take n (unfoldr f s)

Searching by equality

elem :: (Vec v a, Eq a) => a -> v a -> Bool

O(n) elem test if given element is in given vector.

notElem :: (Vec v a, Eq a) => a -> v a -> Bool

O(n) 'not . elem'

elemIndex :: (Vec v a, Eq a) => a -> v a -> Maybe Int

O(n) The elemIndex function returns the index of the first element in the given vector which is equal to the query element, or Nothing if there is no such element.

Misc

data IPair a

Index pair type to help GHC unpack in some loops, useful when write fast folds.

Constructors

IPair !Int a

defaultInitSize :: Int

defaultInitSize = 30, used as initialize size when packing list of unknown size.

chunkOverhead :: Int

The memory management overhead. Currently this is tuned for GHC only.

defaultChunkSize :: Int

The chunk size used for I/O. Currently set to 32k-chunkOverhead

smallChunkSize :: Int

The recommended chunk size. Currently set to 4k - chunkOverhead.

data VectorException

Constructors

IndexOutOfVectorRange !Int CallStack

EmptyVector CallStack

errorEmptyVector :: HasCallStack => a

errorOutRange :: HasCallStack => Int -> a

castVector :: (Vec v a, Cast a b) => v a -> v b

Cast between vectors

C FFI

c_strcmp :: Addr# -> Addr# -> IO CInt

c_strlen :: Addr# -> IO CSize

c_ascii_validate_addr :: Addr# -> Int -> IO Int