Std.IO.Buffered

Description

This module provide buffered I/O interface.

Input & Output device

class Input i where

Input device

Laws: readInput should return 0 on EOF.

Note: readInput is considered not thread-safe, e.g. A Input device can only be used with a single BufferedInput, If multiple BufferedInput s are opened on a same Input device, the behaviour will be undefined.

Methods

readInput :: HasCallStack => i -> Ptr Word8 -> Int -> IO Int

class Output o where

Output device

Laws: writeOutput should not return until all data are written (may not necessarily flushed to hardware, that should be done in device specific way).

Methods

writeOutput :: HasCallStack => o -> Ptr Word8 -> Int -> IO ()

Buffered Input

data BufferedInput i

Input device with buffer, NOT THREAD SAFE!

newBufferedInput

Arguments

:: input

-> Int
Input buffer size

-> IO (BufferedInput input)

readBuffer :: (HasCallStack, Input i) => BufferedInput i -> IO Bytes

Request bytes from BufferedInput.

The buffering logic is quite simple:

If we have pushed back bytes, directly return it, otherwise we read using buffer size. If we read N bytes, and N is larger than half of the buffer size, then we freeze buffer and return, otherwise we copy buffer into result and reuse buffer afterward.

unReadBuffer :: (HasCallStack, Input i) => Bytes -> BufferedInput i -> IO ()

Push bytes back into buffer

readExactly :: (HasCallStack, Input i) => Int -> BufferedInput i -> IO Bytes

Read exactly N bytes

If EOF reached before N bytes read, a ShortReadException will be thrown

readToMagic :: (HasCallStack, Input i) => Word8 -> BufferedInput i -> IO Bytes

Read until reach a magic bytes

If EOF reached before meet a magic byte, the partial bytes are returned.

readToMagic' :: (HasCallStack, Input i) => Word8 -> BufferedInput i -> IO Bytes

Read until reach a magic bytes

If EOF reached before meet a magic byte, a ShortReadException will be thrown.

readLine :: (HasCallStack, Input i) => BufferedInput i -> IO Bytes

Read to a linefeed ('\n' or '\r\n'), return Bytes before it.

If EOF reached before meet a magic byte, the partial bytes are returned.

readLine' :: (HasCallStack, Input i) => BufferedInput i -> IO Bytes

Read to a linefeed ('\n' or '\r\n'), return Bytes before it.

If EOF reached before meet a magic byte, a ShortReadException will be thrown.

Buffered Output

data BufferedOutput o

Output device with buffer, NOT THREAD SAFE!

newBufferedOutput

Arguments

:: output

-> Int
Output buffer size

-> IO (BufferedOutput output)

writeBuffer :: Output o => BufferedOutput o -> Bytes -> IO ()

Write Bytes into buffered handle.

Copy Bytes to buffer if it can hold, otherwise write both buffer(if not empty) and Bytes.

writeBuilder :: Output o => BufferedOutput o -> Builder a -> IO ()

Write Bytes into buffered handle.

Copy Bytes to buffer if it can hold, otherwise write both buffer(if not empty) and Bytes.

flushBuffer :: Output f => BufferedOutput f -> IO ()

Flush the buffer(if not empty).

Exceptions

data ShortReadException

Constructors

ShortReadException IOEInfo