Std.IO.FileSystem

Description

This module provide I/O operations related to filesystem, operations are implemented using unsafe FFIs by default, functions which have T suffix leverage libuv's threadpool to achieve non-block behavior, which should be prefered when the operations' estimated time is long enough(>1ms), such as accessing network filesystem or scan a very large directory. Otherwise you may block RTS's capability thus all the other haskell threads live on it.

regular file devices

data UVFile

data UVFileT

Threaded newtype wrapper for UVFile

data UVFileReader

data UVFileReaderT

data UVFileWriter

data UVFileWriterT

initUVFile

Arguments

:: HasCallStack

=> CBytes

-> UVFileFlag

-> UVFileMode
Sets the file mode (permission and sticky bits), but only if the file was created, see defaultMode.

-> Resource UVFile

Open a file and get the descriptor

initUVFileT

Arguments

:: HasCallStack

=> CBytes

-> UVFileFlag

-> UVFileMode
Sets the file mode (permission and sticky bits), but only if the file was created, see defaultMode.

-> Resource UVFileT

Open a file and get the descriptor

opening constant

data UVFileFlag

uV_FS_O_APPEND :: UVFileFlag

uV_FS_O_CREAT :: UVFileFlag

uV_FS_O_DIRECT :: UVFileFlag

uV_FS_O_DIRECTORY :: UVFileFlag

uV_FS_O_DSYNC :: UVFileFlag

uV_FS_O_EXCL :: UVFileFlag

uV_FS_O_EXLOCK :: UVFileFlag

uV_FS_O_NOATIME :: UVFileFlag

uV_FS_O_NOCTTY :: UVFileFlag

uV_FS_O_NOFOLLOW :: UVFileFlag

uV_FS_O_NONBLOCK :: UVFileFlag

uV_FS_O_RDONLY :: UVFileFlag

uV_FS_O_RDWR :: UVFileFlag

uV_FS_O_SYMLINK :: UVFileFlag

uV_FS_O_SYNC :: UVFileFlag

uV_FS_O_TRUNC :: UVFileFlag

uV_FS_O_WRONLY :: UVFileFlag

uV_FS_O_RANDOM :: UVFileFlag

uV_FS_O_SHORT_LIVED :: UVFileFlag

uV_FS_O_SEQUENTIAL :: UVFileFlag

uV_FS_O_TEMPORARY :: UVFileFlag

type UVFileMode = Int32

defaultMode :: UVFileMode

Default mode for open, 0o666(readable and writable).

filesystem operations

mkdir :: HasCallStack => CBytes -> UVFileMode -> IO ()

mkdirT :: HasCallStack => CBytes -> UVFileMode -> IO ()

unlink :: HasCallStack => CBytes -> IO ()

unlinkT :: HasCallStack => CBytes -> IO ()

mkdtemp :: HasCallStack => CBytes -> IO CBytes

Equivalent to http://linux.die.net/man/3/mkdtemp

Creates a temporary directory in the most secure manner possible. There are no race conditions in the directory’s creation. The directory is readable, writable, and searchable only by the creating user ID. The user of mkdtemp() is responsible for deleting the temporary directory and its contents when done with it.

Note: the argument is the prefix of the temporary directory, so no need to add XXXXXX ending.

mkdtempT :: HasCallStack => CBytes -> IO CBytes