Std.IO.Logger

Description

Simple, high performance logger. The design choice of this logger is biased towards simplicity instead of generlization:

All log functions lives in IO , customization via changeDefaultLogger at the start of the program.

The log format is basically fixed with simple JSON support.

Each logging thread are responsible for building log Builders into a small Bytes with line buffer instead of leaving all Builders to the flushing thread:

We won't keep garbage for too long simply because they're referenced by log's Builder.

Each logging thread only need perform a CAS to prepend log Bytes into a list, which reduces contention.

Logging order is preserved under concurrent settings.

Flushing is automatic and throttled for debug, info, warn to boost performance, while a fatal log always flush logger's buffer, This also lead to a problem that if main thread exits too early logs may missed, to add a flushing when program exits, use withLogger like:

@@@ import Std.IO.Logger

main :: IO () main = withLogger $ do .... @@@

A simple Logger type

data Logger

data LoggerConfig

Logger configuration.

Constructors

LoggerConfig

Fields

loggerBufferSize :: !Int

Buffer size used when creating logger's BufferedOutput

loggerMinFlushInterval :: !Int

Minimal flush interval, see Notes on debug

loggerTsCache :: IO (Builder ())

A IO action return a formatted date/time string

loggerLineBufSize :: !Int

Buffer size to build each log/line

loggerShowDebug :: Bool

Set to False to filter debug logs

loggerShowTS :: Bool

Set to False to disable auto data/time string prepending

newLogger :: Output o => LoggerConfig -> o -> IO Logger

Make a new logger

changeDefaultLogger :: Logger -> IO ()

Change default logger.

withLogger :: IO () -> IO ()

Flush default logger when program exits.

logging functions

debug :: Builder () -> IO ()

info :: Builder () -> IO ()

warn :: Builder () -> IO ()

fatal :: Builder () -> IO ()

otherLevel

Arguments

:: Builder ()
log level

-> Bool
flush immediately?

-> Builder ()
log content

-> IO ()