Std.IO.StdStream

Description

This module provides stdin/stderr/stdout reading and writings. Usually you don't have to use stderr or stderrBuf directly, Logger provides more logging utilities through stderr. While stdinBuf and stdoutBuf is useful when you write interactive programs, Buffered module provide many reading and writing operations. Example:

@@@ import Std.IO.LowResTimer import Std.IO.Buffered import Std.IO.StdStream (stdinBuf, stdoutBuf)

main = do -- read by '\n' b1 <- readLine stdinBuf -- read whatever user input in 3s, otherwise get Nothing b2 <- timeoutLowRes 30 $ readBuffered stdinBuf ... writeBuffer stdoutBuf flushBuffer stdoutBuf

@@@

Standard input & output streams

data StdStream

Standard input and output streams

We support both regular file and TTY based streams, when initialized uv_guess_handle is called to decide which type of devices are connected to standard streams.

StdStream is different from other UVStream in that exception during reading & writing won't close StdStream.

isStdStreamTTY :: StdStream -> Bool

stdin :: StdStream

stdout :: StdStream

stderr :: StdStream

stdinBuf :: BufferedInput StdStream

stdoutBuf :: BufferedOutput StdStream

utils

printStd :: Show a => a -> IO ()

print a Show to stdout

readLineStd :: IO Bytes

read a line from stdin

putStd :: Builder a -> IO ()

print a Builder to stdout

putLineStd :: Builder a -> IO ()

print a Builder to stdout, with a linefeed