Std.IO.UV.FFI

Description

INTERNAL MODULE, provides all libuv side operations.

Documentation

type UVSlot = Int

newtype UVSlotUnSafe

UVSlotUnSafe wrap a slot which may not have a MVar in blocking table, i.e. the blocking table need to be resized.

Constructors

UVSlotUnSafe

Fields

unsafeGetSlot :: UVSlot

type UVFD = Int32

pattern ACCEPT_BUFFER_SIZE :: Int

pattern SO_REUSEPORT_LOAD_BALANCE :: Int

pattern INIT_LOOP_SIZE :: Int

data UVLoop

data UVLoopData

peekUVEventQueue :: Ptr UVLoopData -> IO (Int, Ptr Int)

clearUVEventCounter :: Ptr UVLoopData -> IO ()

peekUVBufferTable :: Ptr UVLoopData -> IO (Ptr (Ptr Word8), Ptr CSsize)

newtype UVRunMode

Constructors

UVRunMode CInt

pattern UV_RUN_DEFAULT :: UVRunMode

pattern UV_RUN_ONCE :: UVRunMode

pattern UV_RUN_NOWAIT :: UVRunMode

peekUVLoopData :: Ptr UVLoop -> IO (Ptr UVLoopData)

Peek loop data pointer from uv loop pointer.

hs_uv_loop_init :: Int -> IO (Ptr UVLoop)

hs_uv_loop_close :: Ptr UVLoop -> IO ()

uv_run :: Ptr UVLoop -> UVRunMode -> IO CInt

uv_run with usafe FFI.

uv_run_safe :: Ptr UVLoop -> UVRunMode -> IO CInt

uv_run with safe FFI.

uv_loop_alive :: Ptr UVLoop -> IO CInt

hs_uv_wake_up_timer :: Ptr UVLoopData -> IO CInt

hs_uv_wake_up_async :: Ptr UVLoopData -> IO CInt

data UVHandle

peekUVHandleData :: Ptr UVHandle -> IO UVSlotUnSafe

hs_uv_fileno :: Ptr UVHandle -> IO UVFD

hs_uv_handle_alloc :: Ptr UVLoop -> IO (Ptr UVHandle)

hs_uv_handle_free :: Ptr UVHandle -> IO ()

hs_uv_handle_close :: Ptr UVHandle -> IO ()

hs_uv_cancel :: Ptr UVLoop -> UVSlot -> IO ()

hs_uv_listen :: Ptr UVHandle -> CInt -> IO CInt

hs_uv_listen_resume :: Ptr UVHandle -> IO ()

hs_uv_read_start :: Ptr UVHandle -> IO CInt

hs_uv_write :: Ptr UVHandle -> Ptr Word8 -> Int -> IO UVSlotUnSafe

hs_uv_accept_check_alloc :: Ptr UVHandle -> IO (Ptr UVHandle)

hs_uv_accept_check_init :: Ptr UVHandle -> IO CInt

hs_uv_accept_check_close :: Ptr UVHandle -> IO ()

hs_uv_tcp_open :: Ptr UVHandle -> UVFD -> IO CInt

uv_tcp_init :: Ptr UVLoop -> Ptr UVHandle -> IO CInt

uv_tcp_init_ex :: Ptr UVLoop -> Ptr UVHandle -> CUInt -> IO CInt

uv_tcp_nodelay :: Ptr UVHandle -> CInt -> IO CInt

uv_tcp_keepalive :: Ptr UVHandle -> CInt -> CUInt -> IO CInt

uV_TCP_IPV6ONLY :: CUInt

uv_tcp_bind :: Ptr UVHandle -> Ptr SockAddr -> CUInt -> IO CInt

hs_uv_tcp_connect :: Ptr UVHandle -> Ptr SockAddr -> IO UVSlotUnSafe

hs_set_socket_reuse :: Ptr UVHandle -> IO CInt

uv_pipe_init :: Ptr UVLoop -> Ptr UVHandle -> CInt -> IO CInt

newtype UVTTYMode

Constructors

UVTTYMode CInt

pattern UV_TTY_MODE_NORMAL :: UVTTYMode

pattern UV_TTY_MODE_RAW :: UVTTYMode

pattern UV_TTY_MODE_IO :: UVTTYMode

uv_tty_init :: Ptr UVLoop -> Ptr UVHandle -> CInt -> IO CInt

newtype UVFileMode

Constructors

UVFileMode CInt

pattern S_IRWXU :: UVFileMode

00700 user (file owner) has read, write and execute permission

pattern S_IRUSR :: UVFileMode

00400 user has read permission

pattern S_IWUSR :: UVFileMode

00200 user has write permission

pattern S_IXUSR :: UVFileMode

00100 user has execute permission

pattern S_IRWXG :: UVFileMode

00070 group has read, write and execute permission

pattern S_IRGRP :: UVFileMode

00040 group has read permission

pattern S_IWGRP :: UVFileMode

00020 group has write permission

pattern S_IXGRP :: UVFileMode

00010 group has execute permission

pattern S_IRWXO :: UVFileMode

00007 others have read, write and execute permission

pattern S_IROTH :: UVFileMode

00004 others have read permission

pattern S_IWOTH :: UVFileMode

00002 others have write permission

pattern S_IXOTH :: UVFileMode

00001 others have execute permission

pattern DEFAULT_MODE :: UVFileMode

Default mode for open, 0o666(readable and writable).

hs_uv_fs_open :: CString -> UVFileFlag -> UVFileMode -> IO UVFD

hs_uv_fs_close :: UVFD -> IO Int

hs_uv_fs_read :: UVFD -> Ptr Word8 -> Int -> Int64 -> IO Int

hs_uv_fs_write :: UVFD -> Ptr Word8 -> Int -> Int64 -> IO Int

hs_uv_fs_unlink :: CString -> IO Int

hs_uv_fs_mkdir :: CString -> UVFileMode -> IO Int

hs_uv_fs_rmdir :: CString -> IO Int

hs_uv_fs_mkdtemp :: CString -> Int -> CString -> IO Int

hs_uv_fs_open_threaded :: CString -> UVFileFlag -> UVFileMode -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_close_threaded :: UVFD -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_read_threaded :: UVFD -> Ptr Word8 -> Int -> Int64 -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_write_threaded :: UVFD -> Ptr Word8 -> Int -> Int64 -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_unlink_threaded :: CString -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_mkdir_threaded :: CString -> UVFileMode -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_rmdir_threaded :: CString -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_mkdtemp_threaded :: CString -> Int -> CString -> Ptr UVLoop -> IO UVSlotUnSafe

newtype UVFileFlag

Constructors

UVFileFlag CInt

pattern O_APPEND :: UVFileFlag

The file is opened in append mode. Before each write, the file offset is positioned at the end of the file.

pattern O_CREAT :: UVFileFlag

The file is created if it does not already exist.

pattern O_DIRECT :: UVFileFlag

File I/O is done directly to and from user-space buffers, which must be aligned. Buffer size and address should be a multiple of the physical sector size of the block device, (DO NOT USE WITH stdio's BufferedIO)

pattern O_DIRECTORY :: UVFileFlag

If the path is not a directory, fail the open. (Not useful on regular file)

Note o_DIRECTORY is not supported on Windows.

pattern O_DSYNC :: UVFileFlag

The file is opened for synchronous I/O. Write operations will complete once all data and a minimum of metadata are flushed to disk.

Note o_DSYNC is supported on Windows via FILE_FLAG_WRITE_THROUGH.

pattern O_EXCL :: UVFileFlag

If the o_CREAT flag is set and the file already exists, fail the open.

Note In general, the behavior of o_EXCL is undefined if it is used without o_CREAT. There is one exception: on Linux 2.6 and later, o_EXCL can be used without o_CREAT if pathname refers to a block device. If the block device is in use by the system (e.g., mounted), the open will fail with the error EBUSY.

pattern O_EXLOCK :: UVFileFlag

Atomically obtain an exclusive lock.

Note UV_FS_O_EXLOCK is only supported on macOS and Windows. (libuv: Changed in version 1.17.0: support is added for Windows.)

pattern O_NOATIME :: UVFileFlag

Do not update the file access time when the file is read.

Note o_NOATIME is not supported on Windows.

pattern O_NOCTTY :: UVFileFlag

If the path identifies a terminal device, opening the path will not cause that terminal to become the controlling terminal for the process (if the process does not already have one). (Not sure if this flag is useful)

Note o_NOCTTY is not supported on Windows.

pattern O_NOFOLLOW :: UVFileFlag

If the path is a symbolic link, fail the open.

Note o_NOFOLLOW is not supported on Windows.

pattern O_NONBLOCK :: UVFileFlag

Open the file in nonblocking mode if possible. (Definitely not useful with stdio)

Note o_NONBLOCK is not supported on Windows. (Not useful on regular file anyway)

pattern O_RANDOM :: UVFileFlag

Access is intended to be random. The system can use this as a hint to optimize file caching.

Note o_RANDOM is only supported on Windows via FILE_FLAG_RANDOM_ACCESS.

pattern O_RDONLY :: UVFileFlag

Open the file for read-only access.

pattern O_RDWR :: UVFileFlag

Open the file for read-write access.

pattern O_SEQUENTIAL :: UVFileFlag

Access is intended to be sequential from beginning to end. The system can use this as a hint to optimize file caching.

Note o_SEQUENTIAL is only supported on Windows via FILE_FLAG_SEQUENTIAL_SCAN.

pattern O_SHORT_LIVED :: UVFileFlag

The file is temporary and should not be flushed to disk if possible.

Note o_SHORT_LIVED is only supported on Windows via FILE_ATTRIBUTE_TEMPORARY.

pattern O_SYMLINK :: UVFileFlag

Open the symbolic link itself rather than the resource it points to.

pattern O_SYNC :: UVFileFlag

The file is opened for synchronous I/O. Write operations will complete once all data and all metadata are flushed to disk.

Note o_SYNC is supported on Windows via FILE_FLAG_WRITE_THROUGH.

pattern O_TEMPORARY :: UVFileFlag

The file is temporary and should not be flushed to disk if possible.

Note o_TEMPORARY is only supported on Windows via FILE_ATTRIBUTE_TEMPORARY.

pattern O_TRUNC :: UVFileFlag

If the file exists and is a regular file, and the file is opened successfully for write access, its length shall be truncated to zero.

pattern O_WRONLY :: UVFileFlag

Open the file for write-only access.

newtype UVDirEntType

Constructors

UVDirEntType CChar

data DirEntType

Constructors

DirEntUnknown

DirEntFile

DirEntDir

DirEntLink

DirEntFIFO

DirEntSocket

DirEntChar

DirEntBlock

fromUVDirEntType :: UVDirEntType -> DirEntType

uV__DT_FILE :: UVDirEntType

uV__DT_DIR :: UVDirEntType

uV__DT_LINK :: UVDirEntType

uV__DT_FIFO :: UVDirEntType

uV__DT_SOCKET :: UVDirEntType

data UVDirEnt

uV__DT_CHAR :: UVDirEntType

peekUVDirEnt :: Ptr UVDirEnt -> IO (CString, UVDirEntType)

uV__DT_BLOCK :: UVDirEntType

hs_uv_fs_scandir_cleanup :: Ptr (Ptr UVDirEnt) -> Int -> IO ()

hs_uv_fs_scandir :: CString -> MBA# (Ptr UVDirEnt) -> IO Int

hs_uv_fs_scandir_extra_cleanup :: Ptr (Ptr (Ptr UVDirEnt)) -> Int -> IO ()

hs_uv_fs_scandir_threaded :: CString -> Ptr (Ptr (Ptr UVDirEnt)) -> Ptr UVLoop -> IO UVSlotUnSafe

data UVTimeSpec

Constructors

UVTimeSpec

Fields

uvtSecond :: !CLong

uvtNanoSecond :: !CLong

data UVStat

Constructors

UVStat

Fields

stDev :: !Word64

stMode :: !Word64

stNlink :: !Word64

stUid :: !Word64

stGid :: !Word64

stRdev :: !Word64

stIno :: !Word64

stSize :: !Word64

stBlksize :: !Word64

stBlocks :: !Word64

stFlags :: !Word64

stGen :: !Word64

stAtim :: !UVTimeSpec

stMtim :: !UVTimeSpec

stCtim :: !UVTimeSpec

stBirthtim :: !UVTimeSpec

uvStatSize :: Int

peekUVStat :: Ptr UVStat -> IO UVStat

hs_uv_fs_stat :: CString -> Ptr UVStat -> IO Int

hs_uv_fs_fstat :: UVFD -> Ptr UVStat -> IO Int

hs_uv_fs_lstat :: CString -> Ptr UVStat -> IO Int

hs_uv_fs_rename :: CString -> CString -> IO Int

hs_uv_fs_fsync :: UVFD -> IO Int

hs_uv_fs_fdatasync :: UVFD -> IO Int

hs_uv_fs_ftruncate :: UVFD -> Int64 -> IO Int

hs_uv_fs_stat_threaded :: CString -> Ptr UVStat -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_fstat_threaded :: UVFD -> Ptr UVStat -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_lstat_threaded :: CString -> Ptr UVStat -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_rename_threaded :: CString -> CString -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_fsync_threaded :: UVFD -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_fdatasync_threaded :: UVFD -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_ftruncate_threaded :: UVFD -> Int64 -> Ptr UVLoop -> IO UVSlotUnSafe

newtype UVCopyFileFlag

Flags control copying.

COPYFILE_EXCL: If present, uv_fs_copyfile() will fail with UV_EEXIST if the destination path already exists. The default behavior is to overwrite the destination if it exists.

COPYFILE_FICLONE: If present, uv_fs_copyfile() will attempt to create a copy-on-write reflink. If the underlying platform does not support copy-on-write, then a fallback copy mechanism is used.

Constructors

UVCopyFileFlag CInt

pattern COPYFILE_DEFAULT :: UVCopyFileFlag

pattern COPYFILE_EXCL :: UVCopyFileFlag

pattern COPYFILE_FICLONE :: UVCopyFileFlag

hs_uv_fs_copyfile :: CString -> CString -> UVCopyFileFlag -> IO Int

hs_uv_fs_copyfile_threaded :: CString -> CString -> UVCopyFileFlag -> Ptr UVLoop -> IO UVSlotUnSafe

newtype UVAccessMode

Constructors

UVAccessMode CInt

pattern F_OK :: UVAccessMode

pattern R_OK :: UVAccessMode

pattern W_OK :: UVAccessMode

pattern X_OK :: UVAccessMode

data AccessResult

Constructors

NoExistence

NoPermission

AccessOK

hs_uv_fs_access :: CString -> UVAccessMode -> IO Int

hs_uv_fs_access_threaded :: CString -> UVAccessMode -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_chmod :: CString -> UVFileMode -> IO Int

hs_uv_fs_chmod_threaded :: CString -> UVFileMode -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_fchmod :: UVFD -> UVFileMode -> IO Int

hs_uv_fs_fchmod_threaded :: UVFD -> UVFileMode -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_utime :: CString -> Double -> Double -> IO Int

hs_uv_fs_utime_threaded :: CString -> Double -> Double -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_futime :: UVFD -> Double -> Double -> IO Int

hs_uv_fs_futime_threaded :: UVFD -> Double -> Double -> Ptr UVLoop -> IO UVSlotUnSafe

newtype UVSymlinkFlag

Constructors

UVSymlinkFlag CInt

pattern SYMLINK_DEFAULT :: UVSymlinkFlag

pattern SYMLINK_DIR :: UVSymlinkFlag

pattern SYMLINK_JUNCTION :: UVSymlinkFlag

hs_uv_fs_link :: CString -> CString -> IO Int

hs_uv_fs_link_threaded :: CString -> CString -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_symlink :: CString -> CString -> UVSymlinkFlag -> IO Int

hs_uv_fs_symlink_threaded :: CString -> CString -> UVSymlinkFlag -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_readlink_cleanup :: CString -> IO ()

hs_uv_fs_readlink :: CString -> MBA# CString -> IO Int

hs_uv_fs_realpath :: CString -> MBA# CString -> IO Int

hs_uv_fs_readlink_extra_cleanup :: Ptr CString -> IO ()

hs_uv_fs_readlink_threaded :: CString -> Ptr CString -> Ptr UVLoop -> IO UVSlotUnSafe

hs_uv_fs_realpath_threaded :: CString -> Ptr CString -> Ptr UVLoop -> IO UVSlotUnSafe

newtype UVHandleType

Constructors

UVHandleType CInt

pattern UV_UNKNOWN_HANDLE :: UVHandleType

pattern UV_ASYNC :: UVHandleType

pattern UV_CHECK :: UVHandleType

pattern UV_FS_EVENT :: UVHandleType

pattern UV_FS_POLL :: UVHandleType

pattern UV_HANDLE :: UVHandleType

pattern UV_IDLE :: UVHandleType

pattern UV_NAMED_PIPE :: UVHandleType

pattern UV_POLL :: UVHandleType

pattern UV_PREPARE :: UVHandleType

pattern UV_PROCESS :: UVHandleType

pattern UV_STREAM :: UVHandleType

pattern UV_TCP :: UVHandleType

pattern UV_TIMER :: UVHandleType

pattern UV_TTY :: UVHandleType

pattern UV_UDP :: UVHandleType

pattern UV_SIGNAL :: UVHandleType

pattern UV_FILE :: UVHandleType

uv_guess_handle :: UVFD -> IO UVHandleType