System.IO.Buffered

Description

This module provide basic I/O interface stdio use.

Documentation

class Input i where

Input device

Convention: input should return 0 on EOF.

Minimal complete definition

readInput

Methods

readInput :: HasCallStack => i -> Ptr Word8 -> Int -> IO Int

class Output o where

Output device

Minimal complete definition

writeOutput

Methods

writeOutput :: HasCallStack => o -> Ptr Word8 -> Int -> IO ()

data BufferedInput i

Constructors

BufferedInput

Fields

bufInput :: i

bufPushBack :: !(IORef Bytes)

inputBuffer :: !(IORef (MutablePrimArray RealWorld Word8))

data BufferedOutput o

Constructors

BufferedOutput

Fields

bufOutput :: o

bufIndex :: !(IORefU Int)

outputBuffer :: !(MutablePrimArray RealWorld Word8)

newBufferedInput :: input -> Int -> IO (BufferedInput input)

newBufferedOutput :: output -> Int -> IO (BufferedOutput output)

readBuffer :: (HasCallStack, Input i) => BufferedInput i -> IO Bytes

Request bytes from BufferedInput.

The buffering logic is quite simple:

If we have pushed back bytes, directly return it, otherwise we read using buffer size. If we read N bytes, and N is larger than half of the buffer size, then we freeze buffer and return, otherwise we copy buffer into result and reuse buffer afterward.

readExactly :: (HasCallStack, Input i) => Int -> BufferedInput i -> IO Bytes

Read exactly N bytes

If EOF reached before N bytes read, a ShortReadException will be thrown

data ShortReadException

Constructors

ShortReadException IOEInfo

unReadBuffer :: (HasCallStack, Input i) => Bytes -> BufferedInput i -> IO ()

Push bytes back into buffer

readToMagic :: (HasCallStack, Input i) => Word8 -> BufferedInput i -> IO Bytes

Read until reach a magic bytes

If EOF reached before meet a magic byte, a ShortReadException will be thrown.

readLine :: (HasCallStack, Input i) => BufferedInput i -> IO Text

Read a line

readTextChunkUTF8 :: (HasCallStack, Input i) => BufferedInput i -> IO Text

Read a chunk of utf8 text

data UTF8DecodeException

Constructors

UTF8InvalidBytesException Bytes IOEInfo

UTF8PartialBytesException Bytes IOEInfo

writeBuffer :: Output o => BufferedOutput o -> Bytes -> IO ()

Write Bytes into buffered handle.

Copy Bytes to buffer if it can hold, otherwise write both buffer(if not empty) and Bytes.

writeBuilder :: Output o => BufferedOutput o -> Builder -> IO ()

Write Bytes into buffered handle.

Copy Bytes to buffer if it can hold, otherwise write both buffer(if not empty) and Bytes.

flush :: Output f => BufferedOutput f -> IO ()

Flush the buffer(if not empty).