System.IO.Exception

Description

This module implemented extensible io exception following approach described in /An Extensible Dynamically-Typed Hierarchy of Exceptions/ by Simon Marlow. The implementation in this module has simplified to meet common need. User who want to catch certain type of exceptions can directly use exception types this module provide, which are modeled after IOErrorType from GHC.IO.Exception.

This module also implements Gabriel Gonzalez'd idea on Resource applicative: http://www.haskellforall.com/2013/06/the-resource-applicative.html. The Applicative and Monad instance is especially useful when you want safely combine multiple resources.

Functions from this package will throw exceptions from this module only instead of the old IOError on I/O exceptions. Exceptions from this module contain IOEInfo which is pretty detailed, but this also require user of this module do some extra work to keep error message's quality(provide CallStack, device informations, etc.). New defined I/O exceptions are encouraged to include a IOEInfo, since it helps a lot when debugging.

Example for library author defining new io exception:

  data MyNetworkException = MyNetworkException IOEInfo ... deriving (Show, Typeable)
  instance Exception MyNetworkException where
        toException = ioExceptionToException
        fromException = ioExceptionFromException

If you're dealing with OS's errno directly, you should convert the errno to libuv's errno in C side with uv_translate_sys_error, then use 'throwUVIfMinus/throwUVError' from this module.

The SomeIOException type

data SomeIOException

The root type of all io exceptions, you can catch all io exception by catching this root type.

Constructors

Exception e => SomeIOException e

ioExceptionToException :: Exception e => e -> SomeException

ioExceptionFromException :: Exception e => SomeException -> Maybe e

Builtin io exception types

data IOEInfo

IO exceptions informations.

Constructors

IOEInfo

Fields

ioeName :: String

the errno name, e.g. EADDRINUSE, etc. empty if no errno.

ioeDescription :: String

description for this io error, can be errno description, or some custom description if no errno.

ioeCallStack :: CallStack

lightweight partial call-stack

data AlreadyExists

Constructors

AlreadyExists IOEInfo

data NoSuchThing

Constructors

NoSuchThing IOEInfo

data ResourceBusy

Constructors

ResourceBusy IOEInfo

data ResourceExhausted

Constructors

ResourceExhausted IOEInfo

data EOF

Constructors

EOF IOEInfo

data IllegalOperation

Constructors

IllegalOperation IOEInfo

data PermissionDenied

Constructors

PermissionDenied IOEInfo

data UnsatisfiedConstraints

Constructors

UnsatisfiedConstraints IOEInfo

data SystemError

Constructors

SystemError IOEInfo

data ProtocolError

Constructors

ProtocolError IOEInfo

data OtherError

Constructors

OtherError IOEInfo

data InvalidArgument

Constructors

InvalidArgument IOEInfo

data InappropriateType

Constructors

InappropriateType IOEInfo

data HardwareFault

Constructors

HardwareFault IOEInfo

data UnsupportedOperation

Constructors

UnsupportedOperation IOEInfo

data TimeExpired

Constructors

TimeExpired IOEInfo

data ResourceVanished

Constructors

ResourceVanished IOEInfo

data Interrupted

Constructors

Interrupted IOEInfo

Throw io exceptions

throwOOMIfNull

Arguments

:: HasCallStack

=> IO (Ptr a)
the allocation action

-> IO (Ptr a)

Throw ResourceExhausted if allocation return a nullPtr.

throwUVIfMinus

Arguments

:: (HasCallStack, Integral a)

=> IO a
the IO action

-> IO a

Throw appropriate IO exception if return value < 0 (libuv's convention).

throwUVIfMinus_

Arguments

:: (HasCallStack, Integral a)

=> IO a
the IO action

-> IO ()

Throw appropriate IO exception if return value < 0, otherwise ignore the result.

Resource management

newtype Resource a

A Resource is an IO action which acquires some resource of type a and also returns a finalizer of type IO () that releases the resource.

The only safe way to use a Resource is with / 'with\'', You should NEVER use the acquire field directly, unless you want to give up resource management.

MonadIO instance is provided so that you can lift IO computation inside Resource, this is convenient for propagate Resource around since many IO computations carry finalizers.

A convention in stdio is that functions returning a Resource should be named in initXXX format, users are strongly recommended to follow this convention.

Constructors

Resource

Fields

acquire :: IO (a, IO ())

initResource :: IO a -> (a -> IO ()) -> Resource a

Create Resource from create and release action.

Note, resource doesn't open resource itself, resource is created when you use with / with'.

initResource_ :: IO () -> IO () -> Resource ()

Create Resource from create and release action.

This function is useful when you want to add some initialization and clean up action inside Resource monad.

withResource :: Resource a -> (a -> IO b) -> IO b

Create a new resource and run some computation, resource is guarantee to be closed.

Be care don't leak the resource through computation return value, because after the computation finishes, the resource is closed already.

withResource' :: Resource a -> (a -> IO () -> IO b) -> IO b

Create a new resource and run some computation, resource is guarantee to be closed.

The difference from with is that the computation will receive an extra close action, which can be used to close the resource early before the whole computation finished, the close action can be called multiple times, only the first call will clean up the resource.

Re-exports

module Control.Exception

type HasCallStack = ?callStack :: CallStack

Request a CallStack.

NOTE: The implicit parameter ?callStack :: CallStack is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: 4.9.0.0

callStack :: HasCallStack -> CallStack

Return the current CallStack.

Does *not* include the call-site of callStack.

Since: 4.9.0.0