System.IO.Net

Description

This module provides an API for creating TCP or IPC servers and clients. IPC Support is implemented with named pipes on Windows, and UNIX domain sockets on other operating systems.

On UNIX, the local domain is also known as the UNIX domain. The path is a filesystem path name. It gets truncated to sizeof(sockaddr_un.sun_path) - 1, which varies on different operating system between 91 and 107 bytes. The typical values are 107 on Linux and 103 on macOS. The path is subject to the same naming conventions and permissions checks as would be done on file creation. It will be visible in the filesystem, and will persist until unlinked.

On Windows, the local domain is implemented using a named pipe. The path must refer to an entry in \?pipe or \.pipe. Any characters are permitted, but the latter may do some processing of pipe names, such as resolving .. sequences. Despite appearances, the pipe name space is flat. Pipes will not persist, they are removed when the last reference to them is closed. Do not forget JavaScript string escaping requires paths to be specified with double-backslashes, such as:

net.createServer().listen( path.join('\\\\?\\pipe', process.cwd(), myctl));

Documentation

initTCPConnection :: HasCallStack => SockAddr -> Maybe SockAddr -> Resource UVStream

data ServerConfig

A TCP/Pipe server configuration

Constructors

Exception e => ServerConfig

Fields

serverAddr :: SockAddr

serverBackLog :: Int

serverWorker :: UVStream -> IO ()

serverWorkerNoDelay :: Bool

serverWorkerHandler :: e -> IO ()

defaultServerConfig :: ServerConfig

A default hello world server on localhost:8888

Test it with main = startServer defaultServerConfig, now try nc -v 127.0.0.1 8888

startServer :: ServerConfig -> IO ()

Start a server

Fork new worker thread upon a new connection.

module System.IO.Net.SockAddr