System.IO.UV.Manager

Description

This module provide I/O manager which bridge libuv's async interface with ghc's light weight thread.

The main procedures for doing event I/O is:

+ Allocate a slot number using allocSlot. + Prepare you I/O buffer and write them to uv loop with pokeBufferTable(both read and write). + Block your thread with a MVar, using getBlockMVar to get it. + Read the result with getResult, for read it's the read bytes number, for write it will be zero. Use throwIfError to guard error situations. + Return the slot back uv manager with freeSlot.

Usually slots are cache in the IO device so that you don't have to allocate new one before each IO operation. Check System.IO.Socket.TCP as an example.

Documentation

data UVManager

getUVManager :: IO UVManager

Get UVManager runing on the same capability.

getBlockMVar :: UVManager -> UVSlot -> IO (MVar ())

Get MVar from blocking table with given slot.

peekBufferTable :: UVManager -> UVSlot -> IO Int

pokeBufferTable :: UVManager -> UVSlot -> Ptr Word8 -> Int -> IO ()

Poke a prepared buffer and size into loop data under given slot.

NOTE, this action is not protected with withUVManager for effcient reason, you should merge this action with other uv action and put them together inside a withUVManager or 'withUVManager\''. for example:

@@ ... withUVManager' uvm $ do pokeBufferTable uvm rslot buf len uvReadStart handle ... @@

withUVManager :: HasCallStack => UVManager -> (Ptr UVLoop -> IO a) -> IO a

Lock an uv mananger, so that we can safely mutate its uv_loop's state.

libuv is not thread safe, use this function to perform any action which will mutate uv_loop's state.

withUVManager' :: HasCallStack => UVManager -> IO a -> IO a

Lock an uv mananger, so that we can safely mutate its uv_loop's state.

Some action did not request uv_loop pointer explicitly, but will mutate uv_loop underhood, for example: uv_read_start. These actions have to be protected by locking the uv_loop.

In fact most of the libuv's functions are not thread safe, so watch out!

handle/request resources

initUVHandle :: HasCallStack => UVHandleType -> (Ptr UVLoop -> Ptr UVHandle -> IO ()) -> UVManager -> Resource (Ptr UVHandle)

Safely lock an uv manager and perform uv_handle initialization.

Initialization an UV handle usually take two step:

allocate an uv_handle struct with proper size

lock a particular uv_loop from a uv manager, and perform custom initialization, such as uv_tcp_init.

And this is what initUVHandle do, all you need to do is to provide the manager you want to hook the handle onto(usually the one on the same capability, i.e. the one obtained by getUVManager), and provide a custom initialization function.

initUVHandleNoSlot :: HasCallStack => UVHandleType -> (Ptr UVLoop -> Ptr UVHandle -> IO ()) -> UVManager -> Resource (Ptr UVHandle)

Safely lock an uv manager and perform uv_handle initialization without allocating a slot.

This function allocate a handle without pre-allocated slot(data field), then perform initialization.

initUVReq :: HasCallStack => UVReqType -> (Ptr UVLoop -> Ptr UVReq -> IO ()) -> UVManager -> Resource (Ptr UVReq)

Safely lock an uv manager and perform uv_req initialization.

initUVReqNoSlot :: HasCallStack => UVReqType -> Resource (Ptr UVReq)

Safely lock an uv manager without allocating a slot.

initUVFS :: HasCallStack => (Ptr UVLoop -> Ptr UVReq -> UVFSCallBack -> IO ()) -> UVManager -> Resource (Ptr UVReq)

Safely lock an uv manager and perform fs operation.

The uv_fs_t request has one more extra clean up step: uv_fs_req_cleanup, thus for perform fs operation with uv_fs_t you should use this function instead of initUVReq.

initUVFSNoSlot :: HasCallStack => (Ptr UVLoop -> Ptr UVReq -> UVFSCallBack -> IO ()) -> Resource (Ptr UVReq)

Safely lock an uv manager and perform fs operation without allocating a slot.

This function will pass nullPtr as `Ptr UVLoop`, nullFunPtr as UVFSCallBack.

data UVStream

A higher level wrappe for uv_stream_t

Constructors

UVStream

Fields

uvsHandle :: !(Ptr UVHandle)

uvsReadSlot :: !UVSlot

uvsWriteReq :: !(Ptr UVReq)

uvsWriteSlot :: !UVSlot

uvsManager :: UVManager

initUVStream :: HasCallStack => UVHandleType -> (Ptr UVLoop -> Ptr UVHandle -> IO ()) -> UVManager -> Resource UVStream

initTCPStream :: HasCallStack => UVManager -> Resource UVStream

initTCPExStream :: HasCallStack => CUInt -> UVManager -> Resource UVStream

initTTYStream :: HasCallStack => UVFD -> UVManager -> Resource UVStream

initPipeStream :: HasCallStack => UVManager -> Resource UVStream

concurrent helpers

forkBa :: IO () -> IO ThreadId

Fork a new GHC thread with active load-balancing.

Using libuv based I/O solution has a disadvantage that file handlers are bound to certain uv_loop, thus certain uv mananger/capability. Worker threads that migrate to other capability may facing contention since various API here is protected by a running lock, this makes GHC's work-stealing strategy unsuitable for certain workload, such as a webserver. we solve this problem with simple round-robin load-balancing: forkBa will automatically distribute your new threads to all capabilities in round-robin manner. Thus its name forkBa(lance).