System.LowResTimer

Description

This module provide low resolution (0.1s) timers using a timing wheel of size 128 per capability, each timer thread will automatically started or stopped based on demannd. register or cancel a timeout is O(1), and each step only need scan n/128 items given timers are registered in an even fashion.

This timer is particularly suitable for high concurrent approximated I/O timeout scheduling. You should not rely on it to provide timing information since it's not very accurate.

Reference:

https://github.com/netty/netty/blob/4.1/common/src/main/java/io/netty/util/HashedWheelTimer.java

http://www.cse.wustl.edu/~cdgill/courses/cs6874/TimingWheels.ppt

low resolution timers

registerLowResTimer

Arguments

:: Int
timout in unit of 100 milliseconds / 0.1s

-> IO ()
the action you want to perform, it should not block

-> IO (IO ())
cancel action

Register a new timer on current capability's timer manager, start the timing wheel if it's not turning.

If the action could block, you may want to run it in another thread. Example to kill a thread after 10s:

  registerLowResTimer 100 (forkIO $ killThread tid)

registerLowResTimerOn

Arguments

:: LowResTimerManager
a low resolution timer manager

-> Int
timout in unit of 100 milliseconds / 0.1s

-> IO ()
the action you want to perform, it should not block

-> IO (IO ())
cancel action

Same as registerLowResTimer, but allow you choose timer manager.

debounce

Arguments

:: Int
cache time in unit of 100 milliseconds / 0.1s

-> IO a
the original IO action

-> IO (IO a)
debounced IO action

Cache result of an IO action for give time t.

This combinator is useful when you want to share IO result within a period, the action will be called on demand, and the result will be cached for t milliseconds.

One common way to get a shared periodical updated value is to start a seperate thread, but doing that will stop system from being idle, which stop idle GC from running, and in turn disable deadlock detection, which is too bad. This function solves that.

low resolution timer manager

data LowResTimerManager

getLowResTimerManager :: IO LowResTimerManager

Get a LowResTimerManager for current thread.

isLowResTimerManagerRunning :: LowResTimerManager -> IO Bool

Check if a timer manager's wheel is turning

This is mostly for testing purpose.

lowResTimerManagerCapabilitiesChanged :: IO ()

Create new low resolution timer manager on capability change.

Since low resolution timer manager is not hooked into RTS, you're responsible to call this function after you call setNumCapabilities to match timer manager array size with new capability number.

This is not a must though, when we fetch timer manager we always take a modulo.