Std.Data.Array.Checked

Description

This module provides exactly the same API with Std.Data.Array, but will throw an IndexOutOfBounds ArrayException on bound check failure, it's useful when debugging array algorithms: just swap this module with Std.Data.Array, segmentation faults caused by out bound access will be turned into exceptions with more informations.

Arr typeclass re-export

class Arr (marr :: * -> * -> *) (arr :: * -> *) a | arr -> marr, marr -> arr

A typeclass to unify box & unboxed, mutable & immutable array operations.

Most of these functions simply wrap their primitive counterpart, if there's no primitive ones, we polyfilled using other operations to get the same semantics.

One exception is that shrinkMutableArr only perform closure resizing on PrimArray because current RTS support only that, shrinkMutableArr will do nothing on other array type.

It's reasonable to trust GHC with specializing & inlining these polymorphric functions. They are used across this package and perform identical to their monomophric counterpart.

Minimal complete definition

newArr, newArrWith, readArr, writeArr, setArr, indexArr, indexArr', indexArrM, freezeArr, thawArr, unsafeFreezeArr, unsafeThawArr, copyArr, copyMutableArr, moveArr, cloneArr, cloneMutableArr, resizeMutableArr, shrinkMutableArr, sameMutableArr, sizeofArr, sizeofMutableArr, sameArr

data RealWorld :: Type

RealWorld is deeply magical. It is primitive, but it is not unlifted (hence ptrArg). We never manipulate values of type RealWorld; it's only used in the type system, to parameterise State#.

Bound checked array operations

newArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => Int -> m (marr s a)

newArrWith :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => Int -> a -> m (marr s a)

readArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => marr s a -> Int -> m a

writeArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => marr s a -> Int -> a -> m ()

setArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => marr s a -> Int -> Int -> a -> m ()

indexArr :: (Arr marr arr a, HasCallStack) => arr a -> Int -> a

indexArr' :: (Arr marr arr a, HasCallStack) => arr a -> Int -> (#a#)

indexArrM :: (Arr marr arr a, Monad m, HasCallStack) => arr a -> Int -> m a

freezeArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => marr s a -> Int -> Int -> m (arr a)

thawArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => arr a -> Int -> Int -> m (marr s a)

copyArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => marr s a -> Int -> arr a -> Int -> Int -> m ()

copyMutableArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => marr s a -> Int -> marr s a -> Int -> Int -> m ()

moveArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => marr s a -> Int -> marr s a -> Int -> Int -> m ()

cloneArr :: (Arr marr arr a, HasCallStack) => arr a -> Int -> Int -> arr a

cloneMutableArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => marr s a -> Int -> Int -> m (marr s a)

resizeMutableArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => marr s a -> Int -> m (marr s a)

shrinkMutableArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s, HasCallStack) => marr s a -> Int -> m ()

New size should be >= 0, and <= original size.

No bound checked operations

unsafeFreezeArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s) => marr s a -> m (arr a)

In place freeze a mutable array, the original mutable array can not be used anymore.

unsafeThawArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s) => arr a -> m (marr s a)

In place thaw a immutable array, the original immutable array can not be used anymore.

sameMutableArr :: Arr marr arr a => marr s a -> marr s a -> Bool

Is two mutable array are reference equal.

sizeofArr :: Arr marr arr a => arr a -> Int

Size of immutable array.

sizeofMutableArr :: (Arr marr arr a, PrimMonad m, PrimState m ~ s) => marr s a -> m Int

Size of mutable array.

sameArr :: Arr marr arr a => arr a -> arr a -> Bool

Is two immutable array are referencing the same one.

Note that sameArr 's result may change depending on compiler's optimizations, for example let arr = runST ... in arr sameArr arr may return false if compiler decides to inline it.

See https://ghc.haskell.org/trac/ghc/ticket/13908 for more background.

Boxed array type

data Array a

Constructors

Array

Fields

array# :: Array# a

data MutableArray s a

Constructors

MutableArray

Fields

marray# :: MutableArray# s a

data SmallArray a

Constructors

SmallArray (SmallArray# a)

data SmallMutableArray s a

Constructors

SmallMutableArray (SmallMutableArray# s a)

uninitialized :: a

Bottom value (throw (UndefinedElement "Data.Array.uninitialized")) for initialize new boxed array(Array, SmallArray..).

NOTE: These functions may segfault when used with indices which are out of bounds.

Primitive array type

data PrimArray a

Constructors

PrimArray ByteArray#

data MutablePrimArray s a

Constructors

MutablePrimArray (MutableByteArray# s)

Bound checked primitive array operations

newPinnedPrimArray :: (PrimMonad m, Prim a, HasCallStack) => Int -> m (MutablePrimArray (PrimState m) a)

Create a pinned byte array of the specified size, The garbage collector is guaranteed not to move it.

newAlignedPinnedPrimArray :: (PrimMonad m, Prim a, HasCallStack) => Int -> m (MutablePrimArray (PrimState m) a)

Create a pinned primitive array of the specified size and respect given primitive type's alignment. The garbage collector is guaranteed not to move it.

copyPrimArrayToPtr :: (PrimMonad m, Prim a, HasCallStack) => Ptr a -> PrimArray a -> Int -> Int -> m ()

copyMutablePrimArrayToPtr :: (PrimMonad m, Prim a, HasCallStack) => Ptr a -> MutablePrimArray (PrimState m) a -> Int -> Int -> m ()

copyPtrToMutablePrimArray :: (PrimMonad m, Prim a, HasCallStack) => MutablePrimArray (PrimState m) a -> Int -> Ptr a -> Int -> m ()

No bound checked primitive array operations

primArrayContents :: PrimArray a -> Ptr a

Yield a pointer to the array's data.

This operation is only safe on pinned primitive arrays allocated by newPinnedPrimArray or newAlignedPinnedPrimArray, and you have to make sure the PrimArray can outlive the Ptr.

mutablePrimArrayContents :: MutablePrimArray s a -> Ptr a

Yield a pointer to the array's data.

This operation is only safe on pinned primitive arrays allocated by newPinnedPrimArray or newAlignedPinnedPrimArray. and you have to make sure the PrimArray can outlive the Ptr.

withPrimArrayContents :: PrimArray a -> (Ptr a -> IO b) -> IO b

Yield a pointer to the array's data and do computation with it.

This operation is only safe on pinned primitive arrays allocated by newPinnedPrimArray or newAlignedPinnedPrimArray.

Don't pass a forever loop to this function, see #14346.

withMutablePrimArrayContents :: MutablePrimArray RealWorld a -> (Ptr a -> IO b) -> IO b

Yield a pointer to the array's data and do computation with it.

This operation is only safe on pinned primitive arrays allocated by newPinnedPrimArray or newAlignedPinnedPrimArray.

Don't pass a forever loop to this function, see #14346.

isPrimArrayPinned :: PrimArray a -> Bool

Check if a primitive array is pinned.

isMutablePrimArrayPinned :: MutablePrimArray s a -> Bool

Check if a mutable primitive array is pinned.

Unlifted array type

data UnliftedArray e

Constructors

UnliftedArray ArrayArray#

data MutableUnliftedArray s e

Constructors

MutableUnliftedArray (MutableArrayArray# s)

class PrimUnlifted a where

Methods

toArrayArray# :: a -> ArrayArray#

fromArrayArray# :: ArrayArray# -> a

The ArrayException type

data ArrayException

Exceptions generated by array operations

Constructors

IndexOutOfBounds String
An attempt was made to index an array outside its declared bounds.

UndefinedElement String
An attempt was made to evaluate an element of an array that had not been initialized.