Std.Data.Builder

Description

If you're users looking for functions turn data into Builder s, check Binary or Textual. This module is intend to Builder builders.

A Builder records a buffer writing function, which can be mappend in O(1) via composition. In stdio a Builder are designed to deal with different AllocateStrategy, it affects how Builder react when writing across buffer boundaries:

When building a short strict Bytes with 'buildBytes/buildByteswith', we do a DoubleBuffer.

When building a large lazy [Bytes] with 'buildBytesList/buildBytesListwith', we do an InsertChunk.

When building and consuming are interlaced with 'buildAndRun/buildAndRunWith', we do an OneShotAction.

Most of the time using combinators from this module to build Builder s is enough, but in case of rolling something shining from the ground, keep an eye on correct AllocateStrategy handling.

Documentation

data AllocateStrategy

AllocateStrategy will decide how each BuildStep proceed when previous buffer is not enough.

Constructors

DoubleBuffer

InsertChunk !Int

OneShotAction (Bytes -> IO ())

data Buffer

Helper type to help ghc unpack

Constructors

Buffer !(MutablePrimArray RealWorld Word8) !Int

type BuildStep = Buffer -> IO [Bytes]

BuilderStep is a function that fill buffer under given conditions.

newtype Builder

Builder is a monoid to help compose BuilderStep. With next BuilderStep continuation, We can do interesting things like perform some action, or interleave the build process.

Constructors

Builder

Fields

runBuilder :: AllocateStrategy -> BuildStep -> BuildStep

append :: Builder -> Builder -> Builder

empty :: Builder

atMost

Arguments

:: Int
size bound

-> (MutablePrimArray (PrimState IO) Word8 -> Int -> IO Int)
the writer which return a new offset for next write

-> Builder

bytes :: Bytes -> Builder

ensureFree :: Int -> Builder

Ensure that there are at least n many elements available.

doubleBuffer :: Int -> BuildStep -> BuildStep

Handle chunk boundary

insertChunk :: Int -> Int -> BuildStep -> BuildStep

oneShotAction :: (Bytes -> IO ()) -> Int -> BuildStep -> BuildStep

buildBytes :: Builder -> Bytes

buildBytesWith :: Int -> Builder -> Bytes

buildBytesList :: Builder -> [Bytes]

buildBytesListWith :: Int -> Int -> Builder -> [Bytes]

buildAndRun :: (Bytes -> IO ()) -> Builder -> IO ()

buildAndRunWith :: Int -> (Bytes -> IO ()) -> Builder -> IO ()