Std.Data.CBytes

Description

This module provide CBytes and many functions working on it, A CBytes is a wrapper for immutable null-terminated string. The main design target of this type is to ease the bridging of C FFI APIs, since most of the unix APIs use null-terminated string. On windows you're encouraged to use a compatibility layer like 'WideCharToMultiByte/MultiByteToWideChar' and keep the same interface, e.g. libuv do this when deal with file paths.

We neither guarantee to store length info, nor support O(1) slice for CBytes: This will defeat the purpose of null-terminated string which is to save memory, We do save the length if it's created on GHC heap though.

It can be used with OverloadedString, literal encoding is UTF-8 with some modifications: NUL char is encoded to 'C0 80', and '\xD800' ~ '\xDFFF' is encoded as a three bytes normal utf-8 codepoint. This is also how ghc compile string literal into binaries, thus we can use rewrite-rules to construct CBytes value in O(1) without wasting runtime heap.

Note most of the unix API is not unicode awared though, you may find a scandir call return a filename which is not proper encoded in any unicode encoding at all. But still, UTF-8 is recommanded to be used everywhere, and we use UTF-8 assumption in various places, such as displaying CBytes and literals encoding above.

Documentation

data CBytes

A efficient wrapper for immutable null-terminated string which can be automatically freed by ghc garbage collector.

Constructors

CBytesOnHeap !(PrimArray Word8)
On heap pinned PrimArray

CBytesLiteral !CString
String literals with static address

create :: HasCallStack => Int -> (CString -> IO Int) -> IO CBytes

Create a CBytes with IO action.

Note: The initial content of the CString is random, it doesn't even have a proper '\NUL' ternimator, user should take the responsibility to do a proper initialization and return the actual length.

append :: CBytes -> CBytes -> CBytes

empty :: CBytes

concat :: [CBytes] -> CBytes

packCBytes :: String -> CBytes

Pack a String into null-terminated CBytes.

data SP2 s

Constructors

SP2 !Int !(MutablePrimArray s Word8)

unpackCBytes :: CBytes -> String

null :: CBytes -> Bool

length :: CBytes -> Int

fromCStringMaybe :: HasCallStack => CString -> IO (Maybe CBytes)

Copy a CString type into a CBytes, return Nothing if the pointer is NULL.

After copying you're free to free the CString 's memory.

fromCString :: HasCallStack => CString -> IO CBytes

Same with fromCString, but throw InvalidArgument when meet a null pointer.

withCBytes :: CBytes -> (CString -> IO a) -> IO a

Pass CBytes to foreign function as a const char*.

c_strcmp :: CString -> CString -> IO CInt

c_strlen :: CString -> IO CSize

c_strcpy :: MutableByteArray# RealWorld -> CString -> IO ()