Std.Data.Text.UTF8Codec

Description

UTF-8 codecs and helpers.

Documentation

encodeCharLength :: Char -> Int

Return a codepoint's encoded length in bytes

If the codepoint is invalid, we return 3(encoded bytes length of replacement char U+FFFD).

encodeChar :: MutablePrimArray s Word8 -> Int -> Char -> ST s Int

Encode a Char into bytes

Write U+FFFD (encoded as EF BF BD 3 bytes) for invalid unicode codepoint. This function assumed there're enough space for encoded bytes, and return the advanced index.

encodeChar# :: MutableByteArray# s -> Int# -> Char# -> State# s -> (#State# s, Int##)

The unboxed version of encodeChar

This function is marked as NOINLINE to reduce code size.

encodeCBytesChar :: PrimMonad m => MutablePrimArray (PrimState m) Word8 -> Int -> Char -> m Int

Encode a Char into bytes with non-standard UTF-8 encoding(Used in Data.CBytes).

'\NUL' is encoded as two bytes C0 80 , '\xD800' ~ '\xDFFF' is encoded as a three bytes normal utf-8 codepoint. This function assumed there're enough space for encoded bytes, and return the advanced index.

encodeCBytesChar# :: MutableByteArray# s -> Int# -> Char# -> State# s -> (#State# s, Int##)

decodeChar :: PrimArray Word8 -> Int -> (#Char, Int#)

Decode a Char from bytes

This function assumed all bytes are UTF-8 encoded, and the index param point to the beginning of a codepoint, the decoded character and the advancing offset are returned.

It's annoying to use unboxed tuple here but we really want GHC to optimize it away.

decodeChar# :: ByteArray# -> Int# -> (#Char#, Int##)

The unboxed version of decodeChar

This function is marked as NOINLINE to reduce code size.

decodeCharLen :: PrimArray Word8 -> Int -> Int

Decode a codepoint's length in bytes

This function assumed all bytes are UTF-8 encoded, and the index param point to the beginning of a codepoint.

decodeCharLen# :: ByteArray# -> Int# -> Int#

The unboxed version of decodeCharLen

This function is marked as NOINLINE to reduce code size.

decodeCharReverse :: PrimArray Word8 -> Int -> (#Char, Int#)

Decode a Char from bytes in rerverse order.

This function assumed all bytes are UTF-8 encoded, and the index param point to the end of a codepoint, the decoded character and the backward advancing offset are returned.

decodeCharReverse# :: ByteArray# -> Int# -> (#Char#, Int##)

The unboxed version of decodeCharReverse

This function is marked as NOINLINE to reduce code size.

decodeCharLenReverse :: PrimArray Word8 -> Int -> Int

Decode a codepoint's length in bytes in reverse order.

This function assumed all bytes are UTF-8 encoded, and the index param point to the end of a codepoint.

decodeCharLenReverse# :: ByteArray# -> Int# -> Int#

The unboxed version of decodeCharLenReverse

This function is marked as NOINLINE to reduce code size.

validateChar :: PrimArray Word8 -> Int -> Int -> Int

Validate if current index point to a valid utf8 codepoint.

If the codepoint is valid, return the utf8 bytes length, otherwise return the negation of offset we should skip(so that a replacing decoder can meet the security rules). If 0# is returned, then you should feed more bytes to continue validation. Note this function guarantee only return 0 when the trailing bytes are partial valid codepoint. That means you can safely replace them with a replacement char or ignore it if no more bytes are to be fed.

reference: https://howardhinnant.github.io/utf_summary.html

validateChar# :: ByteArray# -> Int# -> Int# -> Int#

The unboxed version of decodeCharLenReverse

This function is marked as NOINLINE to reduce code size.

between# :: Word# -> Word# -> Word# -> Bool

isContinueByte# :: Word# -> Bool

chr1# :: Word# -> Char#

chr2# :: Word# -> Word# -> Char#

chr3# :: Word# -> Word# -> Word# -> Char#

chr4# :: Word# -> Word# -> Word# -> Word# -> Char#

copyChar :: Int -> MutablePrimArray s Word8 -> Int -> PrimArray Word8 -> Int -> ST s ()

Unrolled copy loop for copying a utf8-encoded codepoint from source array to target array.