Std.IO.UV.Manager

Description

This module provide I/O manager which bridge libuv's async interface with ghc's light weight thread.

The main procedures for doing event I/O is:

+ Allocate a slot number using allocSlot. + Prepare you I/O buffer and write them to uv loop with pokeBufferTable(both read and write). + Block your thread with a MVar, using getBlockMVar to get it. + Read the result with getResult, for read it's the read bytes number, for write it will be zero. Use throwIfError to guard error situations. + Return the slot back uv manager with freeSlot.

Usually slots are cache in the IO device so that you don't have to allocate new one before each IO operation. Check System.IO.Socket.TCP as an example.

Documentation

data UVManager

getUVManager :: IO UVManager

Get UVManager runing on the same capability.

getBlockMVar :: UVManager -> UVSlot -> IO (MVar Int)

Get MVar from blocking table with given slot.

peekBufferTable :: UVManager -> UVSlot -> IO Int

pokeBufferTable :: UVManager -> UVSlot -> Ptr Word8 -> Int -> IO ()

Poke a prepared buffer and size into loop data under given slot.

NOTE, this action is not protected with 'withUVManager_ for effcient reason, you should merge this action with other uv action and put them together inside a 'withUVManager_ or 'withUVManager\''. for example:

@@ ... withUVManager_ uvm $ do pokeBufferTable uvm slot buf len uvReadStart handle ... @@

withUVManager :: HasCallStack => UVManager -> (Ptr UVLoop -> IO a) -> IO a

Lock an uv mananger, so that we can safely mutate its uv_loop's state.

libuv is not thread safe, use this function to perform any action which will mutate uv_loop's state.

withUVManager_ :: HasCallStack => UVManager -> IO a -> IO a

Lock an uv mananger, so that we can safely mutate its uv_loop's state.

Some action did not request uv_loop pointer explicitly, but will mutate uv_loop underhood, for example: uv_read_start. These actions have to be protected by locking the uv_loop.

In fact most of the libuv's functions are not thread safe, so watch out!

withUVManagerWrap :: HasCallStack => UVManager -> (Ptr UVLoop -> IO UVSlotUnSafe) -> IO Int

handle/request resources

initUVStream :: HasCallStack => (Ptr UVLoop -> Ptr UVHandle -> IO ()) -> UVManager -> Resource UVStream

Safely lock an uv manager and perform uv_handle initialization.

Initialization an UV handle usually take two step:

allocate an uv_handle struct with proper size

lock a particular uv_loop from a uv manager, and perform custom initialization, such as uv_tcp_init.

And this is what initUVHandle do, all you need to do is to provide the manager you want to hook the handle onto(usually the one on the same capability, i.e. the one obtained by getUVManager), and provide a custom initialization function.

data UVStream

Constructors

UVStream

Fields

uvsHandle :: !(Ptr UVHandle)

uvsSlot :: !UVSlot

uvsManager :: UVManager

uvsClosed :: !(IORef Bool)

concurrent helpers

forkBa :: IO () -> IO ThreadId

Fork a new GHC thread with active load-balancing.

Using libuv based I/O solution has a disadvantage that file handlers are bound to certain uv_loop, thus certain uv mananger/capability. Worker threads that migrate to other capability will face contention since various API here is protected by manager's lock, this makes GHC's work-stealing strategy unsuitable for certain workload, such as a webserver. we solve this problem with simple round-robin load-balancing: forkBa will automatically distribute your new threads to all capabilities in round-robin manner. Thus its name forkBa(lance).